Module lib::assert_abs_diff

Assert for comparing absolute differences.

These macros help with collection lengths, such as for strings, arrays, vectors, iterators, and anything that has a typical .len() method.

Compare an absolute difference with an deltaession:

assert_abs_diff_eq!(a, b, delta) ≈ | a - b | = Δ

assert_abs_diff_ne!(a, b, delta) ≈ | a - b | ≠ Δ

assert_abs_diff_lt!(a, b, delta) ≈ | a - b | < Δ

assert_abs_diff_le!(a, b, delta) ≈ | a - b | ≤ Δ

assert_abs_diff_gt!(a, b, delta) ≈ | a - b | > Δ

assert_abs_diff_ge!(a, b, delta) ≈ | a - b | ≥ Δ

Example

use assertables::*;

let a = 10;
let b = 13;
let delta = 3;
assert_abs_diff_eq!(a, b, delta);

Modules

assert_abs_diff_eq

Assert an absolute difference is equal to a delta expression.

assert_abs_diff_ge

Assert an absolute difference is greater than or equal to a delta expression.

assert_abs_diff_gt

Assert an absolute difference is greater than a delta expression.

assert_abs_diff_le

Assert an absolute difference is less than or equal to a delta expression.

assert_abs_diff_lt

Assert an absolute difference is less than a delta expression.

assert_abs_diff_ne

Assert an absolute difference is not equal to a delta expression.