Module lib::assert_command::assert_command_stderr_is_match

Assert a command stderr string is a match to a regex.

If true, return ().

Otherwise, call panic! with a message and the values of the expressions with their debug representations.

Examples

use std::process::Command;
use regex::Regex;

// Return Ok
let mut command = Command::new("bin/printf-stderr");
command.args(["%s", "hello"]);
let matcher = Regex::new(r"ell").unwrap();
assert_command_stderr_is_match!(command, matcher);
//-> ()

// Panic with error message
let result = panic::catch_unwind(|| {
let mut command = Command::new("bin/printf-stderr");
command.args(["%s", "hello"]);
let matcher = Regex::new(r"zzz").unwrap();
assert_command_stderr_is_match!(command, matcher);
//-> panic!
});
assert!(result.is_err());
let actual = result.unwrap_err().downcast::<String>().unwrap().to_string();
let expect = concat!(
    "assertion failed: `assert_command_stderr_is_match!(left_command, right_matcher)`\n",
    "  left_command label: `command`,\n",
    "  left_command debug: `\"bin/printf-stderr\" \"%s\" \"hello\"`,\n",
    " right_matcher label: `matcher`,\n",
    " right_matcher debug: `Regex(\"zzz\")`,\n",
    "                left: `\"hello\"`,\n",
    "               right: `Regex(\"zzz\")`"
);
assert_eq!(actual, expect);

// Panic with custom message
let result = panic::catch_unwind(|| {
let mut command = Command::new("bin/printf-stderr");
let matcher = Regex::new(r"zzz").unwrap();
assert_command_stderr_is_match!(command, matcher, "message");
//-> panic!
});
assert!(result.is_err());
let actual = result.unwrap_err().downcast::<String>().unwrap().to_string();
let expect = "message";
assert_eq!(actual, expect);

Module macros

assert_command_stderr_is_match

assert_command_stderr_is_match_as_result

debug_assert_command_stderr_is_match