Module lib::assert_count

Assert for comparing counts.

These macros help with collection counts, such as for strings, arrays, vectors, iterators, and anything that has a typical .count() method.

Compare a count with another count:

assert_count_eq!(a, b) ≈ a.count() = b.count()

assert_count_ne!(a, b) ≈ a.count() ≠ b.count()

assert_count_lt!(a, b) ≈ a.count() < b.count()

assert_count_le!(a, b) ≈ a.count() ≤ b.count()

assert_count_gt!(a, b) ≈ a.count() > b.count()

assert_count_ge!(a, b) ≈ a.count() ≥ b.count()

Compare a count with an expression:

assert_count_eq_expr!(a, expr) ≈ a.count() = expr

assert_count_ne_expr!(a, expr) ≈ a.count() ≠ expr

assert_count_lt_expr!(a, expr) ≈ a.count() < expr

assert_count_le_expr!(a, expr) ≈ a.count() ≤ expr

assert_count_gt_expr!(a, expr) ≈ a.count() > expr

assert_count_ge_expr!(a, expr) ≈ a.count() ≥ expr

Example

use assertables::*;

let a = "x".chars();
let b = "x".chars();
assert_count_eq!(a, b);

Modules

assert_count_eq

Assert a count is equal to another count.

assert_count_eq_expr

Assert a count is equal to an expression.

assert_count_ge

Assert a count is greater than or equal to another count.

assert_count_ge_expr

Assert a count is greater than or equal to an expression.

assert_count_gt

Assert a count is greater than another count.

assert_count_gt_expr

Assert a count is greater than an expression.

assert_count_le

Assert a count is less than or equal to another count.

assert_count_le_expr

Assert a count is less than or equal to an expression.

assert_count_lt

Assert a count is less than another count.

assert_count_lt_expr

Assert a count is less than an expression.

assert_count_ne

Assert a count is equal to another count.

assert_count_ne_expr

Assert a count is equal to an expression.