Module lib::assert_io_read_to_string

Assert for comparing input/output reader streams.

These macros help with input/output readers, such as file handles, byte arrays, input streams, the trait ::std::io::Read, and anything that implements a method read_to_string() -> String.

Compare a reader with another reader:

assert_io_read_to_string_eq2!(reader1, reader2) ≈ reader1.read_to_string() = reader2.read_to_string()

assert_io_read_to_string_ne2!(reader1, reader2) ≈ reader1.read_to_string() ≠ reader2.read_to_string()

assert_io_read_to_string_lt2!(reader1, reader2) ≈ reader1.read_to_string() < reader2.read_to_string()

assert_io_read_to_string_le2!(reader1, reader2) ≈ reader1.read_to_string() ≤ reader2.read_to_string()

assert_io_read_to_string_gt2!(reader1, reader2) ≈ reader1.read_to_string() > reader2.read_to_string()

assert_io_read_to_string_ge2!(reader1, reader2) ≈ reader1.read_to_string() ≥ reader2.read_to_string()

Compare a reader with an expression:

assert_io_read_to_string_eq!(reader, expr) ≈ reader.read_to_string() = expr

assert_io_read_to_string_ne!(reader, expr) ≈ reader.read_to_string() ≠ expr

assert_io_read_to_string_lt!(reader, expr) ≈ reader.read_to_string() < expr

assert_io_read_to_string_le!(reader, expr) ≈ reader.read_to_string() ≤ expr

assert_io_read_to_string_gt!(reader, expr) ≈ reader.read_to_string() > expr

assert_io_read_to_string_ge!(reader, expr) ≈ reader.read_to_string() ≥ expr

Compare a reader with its contents:

assert_io_read_to_string_contains!(reader, &containee) ≈ reader.read_to_string().contains(containee)

assert_io_read_to_string_is_match!(reader, &matcher) ≈ matcher.is_match(reader.read_to_string())

Example

use assertables::*;
use std::io::Read;

let mut a = "alfa".as_bytes();
let mut b = "alfa".as_bytes();
assert_io_read_to_string_eq2!(a, b);

Modules

assert_io_read_to_string_contains

Assert a ::std::io::Read read_to_string() contains a pattern.

assert_io_read_to_string_eq

Assert a ::std::io::Read read_to_string() value is equal to an expression.

assert_io_read_to_string_eq2

Assert a ::std::io::Read read_to_string() value is equal to another.

assert_io_read_to_string_ge

Assert a ::std::io::Read read_to_string() value is greater than or equal to an expression.

assert_io_read_to_string_ge2

Assert a ::std::io::Read read_to_string() value is greater than or equal to another.

assert_io_read_to_string_gt

Assert a ::std::io::Read read_to_string() value is greater than an expression.

assert_io_read_to_string_gt2

Assert a ::std::io::Read read_to_string() value is greater than another.

assert_io_read_to_string_is_match

Assert a ::std::io::Read read_to_string() is a match to a regex.

assert_io_read_to_string_le

Assert a ::std::io::Read read_to_string() value is less than or equal to an expression.

assert_io_read_to_string_le2

Assert a ::std::io::Read read_to_string() value is less than or equal to another.

assert_io_read_to_string_lt

Assert a ::std::io::Read read_to_string() value is less than an expression.

assert_io_read_to_string_lt2

Assert a ::std::io::Read read_to_string() value is less than another.

assert_io_read_to_string_matches

Assert a ::std::io::read_to_string(path) is a match to a regex.

assert_io_read_to_string_ne

Assert a ::std::io::Read read_to_string() is not equal to an expression.

assert_io_read_to_string_ne2

Assert a ::std::io::Read read_to_string() is not equal to another.