Module lib::assert_iter

Assert for comparing iter collections.

These macros help with comparison of iter parameters, such as two arrays or two vectors. These macros convert each input using the std::iter::Iterator trait.

assert_iter_eq!(collection1, collection2) ≈ iter a = iter b

assert_iter_ne!(collection1, collection2) ≈ iter a ≠ iter b

assert_iter_lt!(collection1, collection2) ≈ iter a < iter b

assert_iter_le!(collection1, collection2) ≈ iter a ≤ iter b

assert_iter_gt!(collection1, collection2) ≈ iter a > iter b

assert_iter_ge!(collection1, collection2) ≈ iter a ≥ iter b

Example

use assertables::*;

let a = [1, 2];
let b = [1, 2];
assert_iter_eq!(&a, &b);

Modules

assert_iter_eq

Assert an iter is equal to another.

assert_iter_ge

Assert an iter is greater than or equal to another.

assert_iter_gt

Assert an iter is greater than another.

assert_iter_le

Assert an iter is less than or equal to another.

assert_iter_lt

Assert an iter is less than another.

assert_iter_ne

Assert an iter is not equal to another.