Module lib::assert_len

Assert macros for comparing lengths.

These macros help with collection lengths, such as for strings, arrays, vectors, iterators, and anything that has a typical .len() method.

Compare a length with another length:

assert_len_eq!(a, b) ≈ a.len() = b.len()

assert_len_ne!(a, b) ≈ a.len() ≠ b.len()

assert_len_lt!(a, b) ≈ a.len() < b.len()

assert_len_le!(a, b) ≈ a.len() ≤ b.len()

assert_len_gt!(a, b) ≈ a.len() > b.len()

assert_len_ge!(a, b) ≈ a.len() ≥ b.len()

Compare a length with an expression:

assert_len_eq_expr!(a, expr) ≈ a.len() = expr

assert_len_ne_expr!(a, expr) ≈ a.len() ≠ expr

assert_len_lt_expr!(a, expr) ≈ a.len() < expr

assert_len_le_expr!(a, expr) ≈ a.len() ≤ expr

assert_len_gt_expr!(a, expr) ≈ a.len() > expr

assert_len_ge_expr!(a, expr) ≈ a.len() ≥ expr

Example

use assertables::*;
 
let a = "x";
let b = "x";
assert_len_eq!(a, b);

Modules

assert_len_eq

Assert a length is equal to another length.

assert_len_eq_expr

Assert a length is equal to an expression.

assert_len_ge

Assert a length is greater than or equal to another length.

assert_len_ge_expr

Assert a length is greater than or equal to an expression.

assert_len_gt

Assert a length is greater than another length.

assert_len_gt_expr

Assert a length is greater than an expression.

assert_len_le

Assert a length is less than or equal to another length.

assert_len_le_expr

Assert a length is less than or equal to an expression.

assert_len_lt

Assert a length is less than another length.

assert_len_lt_expr

Assert a length is less than an expression.

assert_len_ne

Assert a length is equal to another length.

assert_len_ne_expr

Assert a length is equal to an expression.