Module lib::assert_process

Assert for comparing lengths.

These macros help with collection lengths, such as for strings, arrays, vectors, iterators, and anything that has a typical .len() method.

Compare a length with another length:

assert_process_status_code_value_eq!(a, b) ≈ a.len() = b.len()

assert_process_status_code_value_ne!(a, b) ≈ a.len() ≠ b.len()

assert_process_status_code_value_lt!(a, b) ≈ a.len() < b.len()

assert_process_status_code_value_le!(a, b) ≈ a.len() ≤ b.len()

assert_process_status_code_value_gt!(a, b) ≈ a.len() > b.len()

assert_process_status_code_value_ge!(a, b) ≈ a.len() ≥ b.len()

Compare a length with an expression:

assert_process_status_code_value_eq_x!(a, expr) ≈ a.len() = expr

assert_process_status_code_value_ne_x!(a, expr) ≈ a.len() ≠ expr

assert_process_status_code_value_lt_x!(a, expr) ≈ a.len() < expr

assert_process_status_code_value_le_x!(a, expr) ≈ a.len() ≤ expr

assert_process_status_code_value_gt_x!(a, expr) ≈ a.len() > expr

assert_process_status_code_value_ge_x!(a, expr) ≈ a.len() ≥ expr

Example

use assertables::*;
use std::process::Command;

let mut a = Command::new("bin/exit-with-arg"); a.arg("1");
let mut b = Command::new("bin/exit-with-arg"); b.arg("1");
assert_process_status_code_value_eq!(a, b);

Modules

assert_process_status_code_value_eq

Assert a process status code value is equal to another.

assert_process_status_code_value_eq_x

Assert a process status code value is equal to an expression.

assert_process_status_code_value_ge

Assert a process status code value is greater than or equal to another.

assert_process_status_code_value_ge_x

Assert a process status code value is greater than or equal to an expression.

assert_process_status_code_value_gt

Assert a process status code value is greater than another.

assert_process_status_code_value_gt_x

Assert a process status code value is greater than an expression.

assert_process_status_code_value_le

Assert a process status code value is less than or equal to another.

assert_process_status_code_value_le_x

Assert a process status code value is less than or equal to an expression.

assert_process_status_code_value_lt

Assert a process status code value is less than another.

assert_process_status_code_value_lt_x

Assert a process status code value is less than an expression.

assert_process_status_code_value_ne

Assert a process status code value is not equal to another.

assert_process_status_code_value_ne_x

Assert a process status code value is equal to another.