Module lib::assert_some

Assert for Some(_) items.

These macros help compare Some(…) items, such as ::std::Option::Some or similar.

Assert expression is Some:

assert_some!(a) ≈ a is Some

Compare Some(…) to another Some(…):

assert_some_eq2!(a, b) ≈ (a ⇒ Some(a1) ⇒ a1) = (b ⇒ Some(b1) ⇒ b1)

assert_some_ne2!(a, b) ≈ (a ⇒ Some(a1) ⇒ a1) ≠ (b ⇒ Some(b1) ⇒ b1)

Compare Some(…) to an expression:

assert_some_eq!(a, expr) ≈ (a ⇒ Some(a1) ⇒ a1) = b

assert_some_ne!(a, b) ≈ (a ⇒ Some(a1) ⇒ a1) ≠ b

Example

use assertables::*;

let a: Option<i8> = Option::Some(1);
assert_some!(a);

Modules

assert_some

Assert expression is Some.

assert_some_eq

Assert an expression is Some and its value is equal to an expression.

assert_some_eq2

Assert two expressions are Some and their values are equal.

assert_some_ne

Assert an expression is Some and its value is not equal to an expression.

assert_some_ne2

Assert two expressions are Some and their values are not equal.