Module lib::assets

Functions

create_asset

Crée un nouvel actif (asset) en traitant une requête multipart et en interagissant avec la base de données.

delete_asset

Supprime un actif (asset) de la base de données en fonction de son asset_id.

get_asset_from_user_and_asset_id

Récupère les actifs d’un utilisateur spécifié par uuid et un identifiant d’actif spécifique nftoken_id

get_assets

Récupère tous les actifs

get_user_assets

Récupère les actifs d’un utilisateur spécifique

upload_image

Télécharge une image dans le répertoire “uploads” avec un nom de fichier au format asset_id-user_id.extension Seules les extensions .jpeg et .png sont autorisées. Retourne le chemin du fichier en cas de succès.