Function lib::assets::create_asset

pub async fn create_asset(
    __arg0: Extension<Arc<PgPool>>,
    multipart: Multipart,
) -> Result<CollectyResponse, CollectyResponse>

Crée un nouvel actif (asset) en traitant une requête multipart et en interagissant avec la base de données.

Cette fonction permet de créer un nouvel actif en recevant une requête multipart contenant des informations telles que l’identifiant de l’utilisateur, l’adresse XRP, le nom et la description de l’actif, un secret de client et une image. Elle traite ensuite ces informations, crée un NFT via la blockchain XRPL, et stocke les données dans la base de données.

Arguments

pool - Le pool de connexions à la base de données pour interagir avec la base de données PostgreSQL.

multipart - Les données envoyées par le client sous forme de formulaire multipart contenant des informations sur l’actif et l’image.

Retour

Un Result qui contient une réponse CollectyResponse indiquant si l’opération a été réussie ou non.

En cas de succès, une réponse avec le chemin de l’image est renvoyée avec un code HTTP 200 OK.

En cas d’échec, une réponse avec un message d’erreur et le code HTTP approprié est retournée.