Function lib::client::login

pub async fn login(
    username: &String,
    password: &String,
    basic: &String
) -> Result<String, Error>

Login to iFunny and retrive the Bearer token.

This returns an Error enum that includes the captcha url when encountered.

When successful, it returns the bearer.

This must be primed by inputing the information and then waiting ten seconds to try again the first time a new basic token is used.

Examples

use lib::client::login;
use lib::basic::BasicGenerator;
let basic = BasicGenerator::generate();
let username = "username".to_string();
let password = "password".to_string();
login(&username, &password, &basic); 
// Sleep ten seconds
let bearer = login(&username, &password, &basic);