Struct lib::gain_scheduled::GainScheduler

pub struct GainScheduler<Error: Debug, Interface: Channel<Error, Output = f32>, ParamsStore: PidParams<FIXED_POINT>, const REGIONS: usize, const THRESHOLD_MAX: i32, const THRESHOLD_MIN: i32, const TIMESCALE: i32, const FIXED_POINT: u32> { /* private fields */ }

Provides a simple gain scheduled controller.

This controller allows changing parameters based on some signal set via GainScheduler::set_bucket.

Implementations

impl<Error: Debug, Interface: Channel<Error, Output = f32>, ParamsStore: PidParams<FIXED_POINT>, const REGIONS: usize, const THRESHOLD_MAX: i32, const THRESHOLD_MIN: i32, const TIMESCALE: i32, const FIXED_POINT: u32> GainScheduler<Error, Interface, ParamsStore, REGIONS, THRESHOLD_MAX, THRESHOLD_MIN, TIMESCALE, FIXED_POINT>

pub fn new(channel: Interface, params: [ParamsStore; REGIONS]) -> Self

Creates a new controller that sets the output on the Interface using a PID control strategy.

Panics

This function panics if the gain regions are not provided in ascending order.

pub fn follow(&mut self, reference: f32)

Sets the reference value to follow.

pub fn set_bucket(&mut self, bucketer: f32)

Sets the current bucket bucket or schedule to use.

pub fn register_measurement(&mut self, measurement: (f32, u64))

Registers a new timestamped measurement.

pub fn get_gain(&self) -> ParamsStore

Returns the current gain.

pub fn actuate(&mut self) -> Result<ControlInfo<f32>, Error>

Computes the control signal using a PID control strategy.

if successful it returns the expected value and the read value.

Errors

This function returns an error if the Channel fails.

pub fn compute_output(&mut self) -> ControlInfo<f32>

Computes the controller output at the latest timestamped measurement.

Errors

This function returns an error if any of the underlying numeric conversion fail.