Struct lib::graphics::matrix::Matrix

pub struct Matrix { /* fields omitted */ }

A m x n Matrix is represented here

Implementations

impl Matrix

pub fn new(rows: usize, cols: usize, data: Vec<f64>) -> Self

Returns a new row x column Matrix with a vector that contains the data.

Arguments

rows - An unsigned usize int that represents the number of rows in the Matrix

cols - An unsigned usize int that represents the number of columns in the Matrix

data - A vector comprised of floats that is the body of the Matrix

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let vector = vec![0.0, 0.1, 0.2, 0.3];
let matrix = Matrix::new(2, 2, vector);

pub fn get_num_points(&self) -> usize

Returns the number of points (rows) currently in the Matrix.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let vector = vec![0.0, 0.1, 0.2, 0.3];
let matrix = Matrix::new(2, 2, vector);
let num = matrix.get_num_points();

pub fn identity_matrix(size: usize) -> Self

Returns a new N by N identity Matrix.

Arguments

size - An unsigned usize int that represents the size of the identity Matrix

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let ident = Matrix::identity_matrix(4);

pub fn transpose(&self) -> Self

Returns the transpose Matrix of self.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let ident = Matrix::identity_matrix(4);
let transpose = ident.transpose();

pub fn identifize(&mut self)

Makes self an identity Matrix if the matrix is N by N.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let vector = vec![0.0, 0.1, 0.2, 0.3];
let mut matrix = Matrix::new(2, 2, vector);
matrix.identifize();

pub fn fill(&mut self, float: f64)

Fills every element of self.data with a specific float.

Arguments

float - A f64 float that override every element in self.data

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let vector = vec![0.0, 0.1, 0.2, 0.3];
let mut matrix = Matrix::new(2, 2, vector);
matrix.fill(0.0);

pub fn swap_rows(&mut self, row_one: usize, row_two: usize)

Swaps two rows in self.data.

Arguments

row_one - The index of the first row to be swapped.

row_two - The index of the second row to be swapped.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut ident = Matrix::identity_matrix(4);
ident.swap_rows(0, 1);

pub fn get(&self, row: usize, col: usize) -> f64

Returns the corresponding self.data element given a row and column.

Arguments

row - The index of the row of the data point to be accessed

column - The index of the column of the data point to be accessed

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let ident = Matrix::identity_matrix(4);
let num = ident.get(0, 0);

pub fn set(&mut self, row: usize, col: usize, new_point: f64)

Sets the corresponding self.data element a new value given a row and column.

Arguments

row - The index of the row of the data point to be changed

column - The index of the column of the data point to be changed

new_point - The new value to be added

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut ident = Matrix::identity_matrix(4);
ident.set(0, 0, 100.0);

impl Matrix

pub fn iter_row(&self, r: usize) -> impl Iterator<Item = &f64> + '_

Returns a iterator that iterates over a specific row.

Arguments

row - The index of the row to be interated over

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let ident = Matrix::identity_matrix(4);
let iter = ident.iter_row(0);

pub fn iter_col(&self, column: usize) -> impl Iterator<Item = &f64> + '_

Returns a iterator that iterates over a specific column.

Arguments

column - The index of the column to be interated over

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let ident = Matrix::identity_matrix(4);
let iter = ident.iter_col(0);

pub fn iter_col_mut(
&mut self,
column: usize
) -> impl Iterator<Item = &mut f64> + '_

Returns a mutable iterator that iterates over a specific column.

Arguments

column - The index of the column to be interated over

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut ident = Matrix::identity_matrix(4);
let mut iter = ident.iter_col(0);

pub fn iter_by_point(&self) -> ChunksExact<'_, f64>

Returns a iterator that iterates over the Matrix's points.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let ident = Matrix::identity_matrix(4);
let iter = ident.iter_by_point();

pub fn iter_by_point_mut(&mut self) -> ChunksExactMut<'_, f64>

Returns a mutable iterator that iterates over the Matrix's points.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut ident = Matrix::identity_matrix(4);
let mut iter = ident.iter_by_point_mut();

impl Matrix

pub fn add_point(&mut self, x: f64, y: f64, z: f64)

Adds a new point (x, y, z) to a Matrix.

Arguments

x - A f64 float representing the x corrdinate of a point

y - A f64 float representing the y corrdinate of a point

z - A f64 float representing the z corrdinate of a point

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut matrix = Matrix::new(0, 4, Vec::new());
matrix.add_point(0.0, 0.1, 0.2);

pub fn append_point(&mut self, point: &mut Vec<f64>)

Appends a point in the form of a vector to the edge Matrix.

Arguments

point - a mutable vector that has three floats, which will be append to the Matrix

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut matrix = Matrix::new(0, 4, Vec::new());
let mut vector = vec![0.0, 0.1, 0.2];
matrix.append_point(&mut vector);

pub fn add_edge(&mut self, x0: f64, y0: f64, z0: f64, x1: f64, y1: f64, z1: f64)

Adds a new edge to an edge Matrix.

Arguments

x0 - A f64 float representing the start x corrdinate of an edge

y0 - A f64 float representing the start y corrdinate of an edge

z0 - A f64 float representing the start z corrdinate of an edge

x1 - A f64 float representing the end x corrdinate of an edge

y1 - A f64 float representing the end y corrdinate of an edge

z1 - A f64 float representing the end z corrdinate of an edge

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut matrix = Matrix::new(0, 4, Vec::new());
matrix.add_edge(0.0, 0.1, 0.2, 0.3, 0.4, 0.5);

pub fn add_edge_vec(&mut self, edge: Vec<f64>)

Adds a new edge in the form of a f64 vector to an edge Matrix.

Arguments

edge - A vector with six floats representing two points to be added to the edge Matrix

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut matrix = Matrix::new(0, 4, Vec::new());
let vector = vec![0.0, 0.1, 0.2, 0.3, 0.4, 0.5];
matrix.add_edge_vec(vector);

pub fn mult_matrix(&self, other: &Self) -> Self

Returns a the result of multiplying self by another Matrix. Self's columns must be the same size as the other's Matrix's rwos.

Arguments

other - A reference to a Matrix to be multipled with self

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let mut ident1 = Matrix::identity_matrix(4);
let result = ident1.mult_matrix(&Matrix::identity_matrix(4));

pub fn mult_vector(&self, vector: Vec<f64>)

Returns the resulting vecotr when multiplying by self.

Arguments

vector - A mutable vector that will be mutliplied with self.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let ident1 = Matrix::identity_matrix(4);
let mut vector = vec![0.0, 0.1, 0.2, 1.0];
ident1.mult_vector(vector);

impl Matrix

pub fn reflect_yz() -> Self

Returns a reflection over the yz (y) axis transformation Matrix.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let reflect = Matrix::reflect_yz();

pub fn reflect_xz() -> Self

Returns a reflection over the xz (x) axis transformation Matrix.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let reflect = Matrix::reflect_xz();

pub fn reflect_xy() -> Self

Returns a reflection over the xy (y) axis transformation Matrix.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let reflect = Matrix::reflect_xy();

pub fn reflect_45() -> Self

Returns a reflection over the y=x axis transformation Matrix.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let reflect = Matrix::reflect_45();

pub fn reflect_neg45() -> Self

Returns a reflection over the y=-x axis transformation Matrix.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let reflect = Matrix::reflect_neg45();

pub fn reflect_origin() -> Self

Returns a reflection over the origin transformation Matrix.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let reflect = Matrix::reflect_origin();

pub fn translate(x: f64, y: f64, z: f64) -> Self

Returns a translation transformation Matrix.

Arguments

x - A f64 float specifying the x direction to move.

y - A f64 float specifying the y direction to move.

z - A f64 float specifying the z direction to move.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let translate = Matrix::translate(50.0, -100.0, 0.0);

pub fn scale(x: f64, y: f64, z: f64) -> Self

Returns a dilation transformation Matrix.

Arguments

x - A f64 float specifying the x scale factor.

y - A f64 float specifying the y scale factor.

z - A f64 float specifying the z scale factor.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let dilation = Matrix::scale(0.5, 0.25, 0.1);

pub fn rotate_point(theta: f64, x: f64, y: f64, z: f64) -> Self

Returns a transformation Matrix to allow a vector to be rotated around any axis.

Arguments

theta - A f64 float representing the angle of rotation.

x - A f64 float specifying the x componet of the specified axis

y - A f64 float specifying the y componet of the specified axis

z - A f64 float specifying the z componet of the specified axis

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let rotate = Matrix::rotate_point(180.0, 100.0, 100.0, 100.0);

pub fn rotate_x(theta: f64) -> Self

Returns a rotation over the x axis transformation Matrix.

Arguments

theta - A f64 float representing the angle of rotation.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let rotate = Matrix::rotate_x(45.0);

pub fn rotate_y(theta: f64) -> Self

Returns a rotation over the y axis transformation Matrix.

Arguments

theta - A f64 float representing the angle of rotation.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let rotate = Matrix::rotate_y(45.0);

pub fn rotate_z(theta: f64) -> Self

Returns a rotation over the z axis transformation Matrix.

Arguments

theta - A f64 float representing the angle of rotation.

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let rotate = Matrix::rotate_z(45.0);

pub fn shearing_x(sh_y: f64, sh_z: f64) -> Self

Returns a shearing over the x axis transformation Matrix.

Arguments

sh_y - The factor for which the y axis will shear

sh_z - The factor for which the z axis will shear

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let rotate = Matrix::shearing_x(1.3, 0.5);

pub fn shearing_y(sh_x: f64, sh_z: f64) -> Self

Returns a shearing over the y axis transformation Matrix.

Arguments

sh_x - The factor for which the x axis will shear

sh_z - The factor for which the z axis will shear

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let rotate = Matrix::shearing_y(1.3, 0.5);

pub fn shearing_z(sh_x: f64, sh_y: f64) -> Self

Returns a shearing over the z axis transformation Matrix.

Arguments

sh_x - The factor for which the x axis will shear

sh_y - The factor for which the y axis will shear

Examples

Basic usage:

use transform_rs::graphics::matrix::Matrix;
let rotate = Matrix::shearing_z(1.3, 0.5);