Macro lib::assert_fs_read_to_string_gt_as_result

macro_rules! assert_fs_read_to_string_gt_as_result {
    ($a_path:expr, $b_path:expr $(,)?) => { ... };
}

Assert a std::fs::read_to_string() value is greater than another.

If true, return Result Ok(()).

Otherwise, return Result Err with a diagnostic message.

This macro provides the same statements as assert_fs_read_to_string_gt, except this macro returns a Result, rather than doing a panic.

This macro is useful for runtime checks, such as checking parameters, or sanitizing inputs, or handling different results in different ways.

Module macros

assert_fs_read_to_string_gt

assert_fs_read_to_string_gt_as_result

debug_assert_fs_read_to_string_gt