Macro lib::assert_program_args_stderr_le_expr_as_result

macro_rules! assert_program_args_stderr_le_expr_as_result {
    ($a_program:expr, $a_args:expr, $b_expr:expr $(,)?) => { ... };
}

Assert a command (built with program and args) stderr string is less than or equal to an expression.

If true, return Result Ok(()).

Otherwise, return Result Err with a diagnostic message.

This macro provides the same statements as [assert_], except this macro returns a Result, rather than doing a panic.

This macro is useful for runtime checks, such as checking parameters, or sanitizing inputs, or handling different results in different ways.

Related

[assert_program_args_stderr_le_expr]

[assert_program_args_stderr_le_expr_as_result]

[debug_assert_program_args_stderr_le_expr]