Struct lib::parser::Parser

pub struct Parser { /* fields omitted */ }

Goes through the file named filename and performs all of the actions listed in that file.
The file follows the following format:
     Every command is a single character that takes up a line
     Any command that requires arguments must have those arguments
     in the second line. The commands are as follows: 
         line: add a line to the edge matrix -
               takes 6 arguemnts (x0, y0, z0, x1, y1, z1)
         ident: set the transform matrix to the identity matrix -
         scale: create a scale matrix,
                then multiply the transform matrix by the scale matrix -
                takes 3 arguments (sx, sy, sz)
         move: create a translation matrix,
               then multiply the transform matrix by the translation matrix -
               takes 3 arguments (tx, ty, tz)
         rotate: create a rotation matrix,
                 then multiply the transform matrix by the rotation matrix -
                 takes 2 arguments (axis, theta) axis should be x y or z
         reflect: create a reflection matrix, 
                 then multiply the transform matrix by the rotation matrix -
                 takes a argument (axis) - should be x y or z
         apply: apply the current transformation matrix to the edge matrix
         display: clear the screen, then
                  draw the lines of the edge matrix to the screen
                  display the screen
         save: clear the screen, then
               draw the lines of the edge matrix to the screen
               save the screen to a file -
               takes 1 argument (file name)
         quit: end parsing

Implementations

impl Parser

pub fn new(
file_name: &str,
width: u32,
height: u32,
range: u8,
color: Pixel
) -> Self

Returns a parser that can parse through file_name

Arguments

file_name - The name of the file that will be created.

height - An unsigned int that will represent height of the Canvas

width - An unsigned int that will represent width of the Canvas

range - An unsigned int that will represent maximum depth of colors in the Canvas

color - A Pixel that represents the color of the drawing line

Examples

Basic usage:

use transform_rs::graphics::display::Pixel;
use transform_rs::parser::Parser;
let purplish = Pixel::new(17, 46, 81);
let porygon = Parser::new("./porygon_script", 512, 512, 255, purplish);

pub fn new_with_bg(
file_name: &str,
width: u32,
height: u32,
range: u8,
color: Pixel,
bg: Pixel
) -> Self

Returns a parser that can parse through file_name that starts with Canvas filled by bg.

Arguments

file_name - The name of the file that will be created.

height - An unsigned int that will represent height of the Canvas

height - An unsigned int that will represent height of the Canvas

width - An unsigned int that will represent width of the Canvas

range - An unsigned int that will represent maximum depth of colors in the Canvas

color - A Pixel that represents the color of the drawing line

bg - A Pixel the default background color of self.canvas

Examples

Basic usage:

use transform_rs::graphics::display::Pixel;
use transform_rs::parser::Parser;
let purplish = Pixel::new(17, 46, 81);
let outline = Pixel::new(235, 219, 178);
let porygon = Parser::new_with_bg("./porygon_script", 512, 512, 255, purplish, outline);

pub fn parse_file(&mut self)

Parses and runs through the commands in self.file_name

Examples

Basic usage:

use transform_rs::graphics::display::Pixel;
use transform_rs::parser::Parser;
let purplish = Pixel::new(17, 46, 81);
let outline = Pixel::new(235, 219, 178);
let mut porygon = Parser::new_with_bg("./porygon_script", 512, 512, 255, purplish, outline);
porygon.parse_file();