Macro lib::pid::prelude::new_pid

macro_rules! new_pid {
    (
        buffer_size:
        $buffer:literal,kp:
        $kp:literal,ki:
        $ki:literal,kd:
        $kd:literal,sample_time:
        $ts:literal,threshold_max:
        $max:literal,threshold_min:
        $min:literal,time_scale:
        $time_scale:literal,fixed_point:
        $fixed_point:literal,output:
        $channel:tt
    ) => { ... };
    (
        kp:
        $kp:literal,ki:
        $ki:literal,kd:
        $kd:literal,sample_time:
        $ts:literal,threshold_max:
        $max:literal,threshold_min:
        $min:literal,time_scale:
        $time_scale:literal,fixed_point:
        $fixed_point:literal,output:
        $channel:tt
    ) => { ... };
}

Makes the instantiation of a new PID controller a bit more readable.

Syntax :

  use lib::pid::prelude::*;
  // Note that the channel here is probably some output thing, esc, servo etc.
  let channel = 0f32;
  let target = [6f32, 6f32];
  let mut pid = new_pid!(
    buffer_size:2, // This is optional, leaving it out will generate a buffer of size 1.
    kp:2,
    ki:1,
    kd:1,
    sample_time:1,
    threshold_max:100,
    threshold_min:0,
    time_scale:1,
    fixed_point:0,
    output:channel
  );
  pid.follow(target);

  pid.register_measurement(1f32, 0);
  let actuation = pid.actuate().unwrap();
  let expected = 2f32 * (actuation.reference - actuation.measured) + 5f32 / 2f32 + 5f32;
  assert!(actuation.actuation == expected);

  pid.register_measurement(2f32, 1);
  let actuation = pid.actuate().unwrap();
  let expected =
    2f32 * (actuation.reference - actuation.measured) + (5f32 + 4f32) / 2f32 + 2.5f32 - 1f32;
  // Accumulated sum + average from previous time step to the current time step.
  assert!(actuation.actuation == expected);