Trait lib::pid::DoubleSize

pub trait DoubleSize {
    type Ret: Sized;
    type Error: Debug;

    // Required methods
    fn double_size(self) -> Self::Ret;
    fn half_size(value: Self::Ret) -> Result<Self, Self::Error>
       where Self: Sized;
}

Represents that a type can be doubled in size.

This is trivially true for all of the types it is implemented on.

Required Associated Types

type Ret: Sized

A type with double the size of self.

type Error: Debug

Potential errors that can occur.

Required Methods

fn double_size(self) -> Self::Ret

Doubles the size of self.

fn half_size(value: Self::Ret) -> Result<Self, Self::Error>
where Self: Sized,

Halves the size of the value.

Errors

Returns error if the value does not fit.