lib::raytrace
Trait Raytrace

pub trait Raytrace<T: Copy + Add<Output = T> + Mul<Output = T> + Div<Output = T> + Sub<Output = T>> {
    // Required methods
    fn intersects_along(&self, ray: &Ray<T>) -> Option<T>;
    fn transmit(&self, ray: &Ray<T>) -> Option<Ray<T>>;

    // Provided method
    fn intersects_at(&self, ray: &Ray<T>) -> Option<Vec3<T>> { ... }
}

The trait for all objects within the raytracer.

This trait must be implemented for every object within the scene.

Required Methods

fn intersects_along(&self, ray: &Ray<T>) -> Option<T>

Gives the distance at which a ray intersects the object.

fn transmit(&self, ray: &Ray<T>) -> Option<Ray<T>>

Reflects, refracts, or otherwise transforms the ray in accordance to how the object should behave.

Provided Methods

fn intersects_at(&self, ray: &Ray<T>) -> Option<Vec3<T>>

Gives the position of intersection between a ray and an object.