IRON

Trait lib::response::modifiers::Bodyable

pub trait Bodyable {
    fn set_body(self, res: &mut Response);
}

Something that can be used to set the body of a response.

Required Methods

fn set_body(self, res: &mut Response)

Set the body of this response, possibly also setting headers.