Function lib::runtime::runtime_bindings::halide_memoization_cache_release

pub unsafe extern "C" fn halide_memoization_cache_release(
    user_context: *mut c_void, 
    host: *mut c_void
)

If halide_memoization_cache_lookup succeeds, halide_memoization_cache_release must be called to signal the storage is no longer being used by the caller.

It will be passed to the host pointer of one the buffers returned by halide_memoization_cache_lookup.

That is halide_memoization_cache_release will be called multiple times for the case where halide_memoization_cache_lookup is handling multiple buffers. (This corresponds to memoizing a Tuple in Halide.)

Note that the host pointer must be sufficient to get to all information the release operation needs.

The default Halide cache implementation accomplishes this by storing extra data before the start of the user modifiable host storage.

This call is like free and does not have a failure return.