Struct lib::shareable::ShareableOrderedVec

pub struct ShareableOrderedVec<T> { /* fields omitted */ }

A collection that keeps the ordering of its elements, even when deleting an element However, this collection can be shared between threads We can guess what the index is for an element that we must add We can use get, and get_next_id_increment on other threads, but that is all We must do the rest of our operations using an external messaging system

Implementations

impl<T> ShareableOrderedVec<T>

Actual code that we will update on the main thread, or the creation thread

pub fn insert(&mut self, idx: usize, elem: T) -> Option<T>

Add an element to the ordered vector, but at a specific location This will return the last element that was at that position, if possible

pub fn get_next_idx(&self) -> usize

Get the index of the next element that we will add. If we call this twice, without inserting any elements, it will not change

pub fn get_next_id_increment(&self) -> usize

Check the next ID where we can add an element, but also increment the counter, so it won’t be the same ID

pub fn remove(&mut self, idx: usize) -> Option<T>

Remove an element that was already added

pub fn init_update(&self)

Update the atomic counters at the start, before we do anything on the other threads.

pub fn finish_update(&self)

Update the rest of the stuff at the end, after we edit the Shareable data on the other threads. This should be ran before we run any external messages that were sent by other threads

pub fn get(&self, idx: usize) -> Option<&T>

Get a reference to an element in the ordered vector

pub fn get_mut(&mut self, idx: usize) -> Option<&mut T>

Get a mutable reference to an element in the ordered vector

pub fn count(&self) -> usize

Get the number of valid elements in the ordered vector

pub fn count_invalid(&self) -> usize

Get the number of invalid elements in the ordered vector

pub fn clear(&mut self) -> Vec<Option<T>>

Clear the whole ordered vector

impl<T> ShareableOrderedVec<T>

Iter magic

pub fn iter(&self) -> impl Iterator<Item = &T>

Get an iterator over the valid elements

pub fn iter_indexed(&self) -> impl Iterator<Item = (usize, &T)>

Get an iterator over the valid elements

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut T>

Get a mutable iterator over the valid elements

pub fn iter_indexed_mut(&mut self) -> impl Iterator<Item = (usize, &mut T)>

Get a mutable iterator over the valid elements with their index

pub fn iter_invalid(&self) -> impl Iterator<Item = usize>

Get an iterator over the indices of the null elements

pub fn my_drain<F>(
&mut self,
filter: F
) -> impl Iterator<Item = (usize, T)> + '_ where
F: FnMut(usize, &T) -> bool,

Drain the elements that only return true. This will return just an Iterator of the index and value of the drained elements