Struct lib::DenseBitSet

pub struct DenseBitSet { /* fields omitted */ }

A dense bit set implemented over Vec

Methods

impl DenseBitSet

pub fn with_capacity(num_bits: usize) -> DenseBitSet

Creates a DenseBitSet that can contain at least num_bits bits

Examples

Basic usage:

use bitsets::DenseBitSet;
 

pub fn set(&mut self, i: usize) -> bool

Sets the ith bit

pub fn test(&self, i: usize) -> bool

Tests whether the ith bit is set

pub fn toggle(&mut self, i: usize)

if the ith

pub fn words(&self) -> usize

returns the number of elements in the underlying Vec

pub fn len(&self) -> usize

returns the number of bits this set can accomodate

pub fn print(&self)

prints the bitset to STDOUT

pub fn not(&mut self)

bitwise not all the bits in the bitset

pub fn and(&mut self, other: &DenseBitSet)

Combining functions

The following methods take another bitset, and merge the two together

pub fn or(&mut self, other: &DenseBitSet)

pub fn xor(&mut self, other: &DenseBitSet)