lib

Struct Game

pub struct Game {
    pub fen: String,
    pub board: Board,
    pub checks: Vec<bool>,
    pub game_status: u8,
}

A struct to represent the chess game itself.

Attributes

fen: String, // the FEN string that represents the current position.
board: Board, // A representation of the FEN string that is easier to work with.
checks: Vec<bool>, // index 0 is white's check status, index 1 is black's check status
game_status: u8, // 0: Game in progress, 1: Checkmate (White wins), 2: Checkmate (Black wins), 3: Stalemate, 4: Draw by 50 move rule
 

Fields
fen: Stringboard: Boardchecks: Vec<bool>game_status: u8
Implementations

impl Game

pub fn new_from_fen(fen: String) -> Game

A function to create a new Game object from a given FEN.

Arguments

fen: String, // A valid FEN string.
 

Returns

This function returns a Game object. This function will not error, but it will exhibit undocumented behaviour if the inputted FEN string is invalid.

Example

let example_game = Game::new_from_fen("r1bk3r/p2pBpNp/n4n2/1p1NP2P/6P1/3P4/P1P1K3/q5b1 b - - 1 23".to_string()); // Create a new game at the final position of the Immortal Game. This Game will have status 1, since White is victorious.

pub fn new() -> Game

A function to create a new Game at the starting position. Alias to

Game::new_from_fen("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1".to_string())

Returns

This function returns a valid in-progress Game object.

Example

let example_game = Game::new(); // Create a new game at the starting position

pub fn make_move(&mut self, source: String, target: String) -> bool

A function to make a move in the Game. Automatically detects whose turn it is based on Game.board.

Arguments

source: String, // The square where the piece to move stands, in algebraic notation.
target: String, // The square to which to move the piece, in algebraic notation.
 

Returns

This function returns true if the move was successfully made, false otherwist,

Example

let example_game = Game::new(); // Create a new game at the starting position
example_game.make_move("f2".to_string(), "f3".to_string()); // 1. f3
example_game.make_move("e7".to_string(), "e5".to_string()); // 1... e5
example_game.make_move("g2".to_string(), "g4".to_string()); // 2. g4
example_game.make_move("d8".to_string(), "h4".to_string()); // 2... Qh4#