Struct lib::MerkleTree

pub struct MerkleTree<T> { /* fields omitted */ }

A Merkle tree is a binary tree that has two main types of nodes:

leafs that contain a value (usually block) and it’s hash;

nodes that contain a hash of a left and right nodes hashes concatenation; h0123 /
h01 h23 / \ /
h0 h1 h2 h3

Implementations

impl<T: Clone> MerkleTree<T>

pub fn build_tree(values: Vec<T>) -> Self where
T: Hash,

Constructs a Merkle Tree from a vector of data blocks. Returns None if values is empty.

pub fn root_hash(&self) -> u64

Returns the root hash of Merkle tree

pub fn height(&self) -> usize

Returns the height of Merkle tree

pub fn count(&self) -> usize

Returns the number of leaves in the Merkle tree

pub fn is_empty(&self) -> bool

Returns whether the Merkle tree is empty or not

pub fn gen_proof(&self, value: T) -> Option<Proof<T>> where
T: Hash + PartialEq,

Generate an inclusion proof for the given value. Returns None if the given value is not found in the tree.

pub fn gen_nth_proof(&self, n: usize) -> Option<Proof<T>>

Generate an inclusion proof for the n-th leaf value.

pub fn validate_proof(&self, proof_to_verify: &Proof<T>) -> bool where
T: Hash + PartialEq,

Returns true if proof is valid, false otherwise.