Struct lib::Sudoku

pub struct Sudoku { /* private fields */ }

Sudoku. It contains the data of the field and counts the moves

Implementations

impl Sudoku

pub const UNSOLVED_SUDOKU_FILES: [&str; 8] = _

Relative Path of solveable Sudokus. The following format is used: {website}-{date of retival}-{level}-{solved/unsolved}

For data structure see Sudoku::from_file

pub fn empty() -> Self

Returns an empty Sudoku

use sudoku::Sudoku;

let sudoku = Sudoku::empty();

pub fn example_solved() -> Self

Returns an solved Sudoku

use sudoku::Sudoku;

let sudoku = Sudoku::example_solved();

pub fn from_file(file_path: &str) -> Self

Returns an Sudoku from File

use sudoku::Sudoku;

let sudoku = Sudoku::from_file("test-sudokus/text/sudoku-com-12-11-12-master-solved");

pub fn count_moves(&self) -> u32

Returns the number of already taken moves

use sudoku::Sudoku;

let mut sudoku = Sudoku::from_file("test-sudokus/text/sudoku-com-12-11-12-master-unsolved");
sudoku.solve();
assert!(sudoku.count_moves() > 0);

pub fn is_set(&self, row: usize, collum: usize) -> bool

Returns true, if the field is set. A single Number is writen.

use sudoku::Sudoku;

let sudoku = Sudoku::example_solved();
println!("Left upper Corner is set: {}", sudoku.is_set(0, 0));

pub fn count_unset(&self) -> u32

Counts the number of cells in the sudoku that are not set

use sudoku::Sudoku;

let sudoku = Sudoku::example_solved();
 
// A solved sudoku doesn't have any unset cells
assert_eq!(sudoku.count_unset(), 0);

pub fn clear(&mut self, row: usize, collum: usize)

Clears a cell

use sudoku::Sudoku;

let mut sudoku = Sudoku::example_solved();
 
// Clear the first cell
sudoku.clear(0,0);
 
assert_eq!(sudoku.get_number(0,0), 0);

pub fn set_number(&mut self, row: usize, collum: usize, value: i32)

Sets the number of cell

ToDo: Es ist mehr ein add

use sudoku::Sudoku;

let mut sudoku = Sudoku::example_solved();
 
// Clear the first cell
sudoku.clear(0,0);
 
// Set it to 9
sudoku.set_number(0, 0, 9);
 
assert_eq!(sudoku.get_number(0,0), 9);

pub fn solve(&mut self) -> bool

Solves the sudoku and returns true if it is solvable

use sudoku::Sudoku;

let mut sudoku = Sudoku::from_file("test-sudokus/text/sudoku-com-12-11-12-master-unsolved");
sudoku.solve();
assert!(sudoku.is_correct());

pub fn remove_number(&mut self, row: usize, collum: usize, value: i32)

Removes a number from a field

pub fn get_number(&self, row: usize, collum: usize) -> i32

Returns the number of a cell If no or multible numbers are set it returns 0

pub fn get_numbers(&self, row: usize, collum: usize) -> Vec<i32>

Gets all numbers that are set in a cell as a Vec

pub fn is_solved(&self) -> bool

Returns true if all numbers are set It doesnt have to be solved correctly

use sudoku::Sudoku;

let sudoku = Sudoku::example_solved();
assert!(sudoku.is_solved());

pub fn is_correct(&self) -> bool

Returns true if a sudoku is correctly solved

use sudoku::Sudoku;

let sudoku = Sudoku::example_solved();
assert!(sudoku.is_correct());
 
//The following is always true
assert!(sudoku.is_solved() && sudoku.is_part_correct() == sudoku.is_correct());

pub fn is_part_correct(&self) -> bool

Checks if their are errors in the sudoku field Boxes may be unset and it will return true

use sudoku::Sudoku;

let sudoku = Sudoku::example_solved();
assert!(sudoku.is_part_correct());

pub fn set_missing_numbers(&mut self) -> bool

Finds and sets a missing number in the sudoku field ToDo: Umbennen ung beschreiben, was eine missing number ist

use sudoku::Sudoku;

let mut sudoku = Sudoku::from_file("test-sudokus/text/tagesspiegel-23-11-11-easy-few-missing");
sudoku.set_missing_numbers();

pub fn set_alone_number(&mut self) -> bool

Sets the first alone number A alone number is a number that only appears once in a row, collum or 3x3 square in the possibilitys

The function returns true, if it was able to set a alone number

use sudoku::Sudoku;

let mut sudoku = Sudoku::from_file("test-sudokus/text/tagesspiegel-23-11-11-easy-few-missing");
sudoku.set_alone_number();