Enum lib::types::Message

pub enum Message<T: HasName + HasTarget> {
    Quit,
    Event(T),
    Busy(String),
    Free(String),
}

Message type used as container for all Supevisors/Workers communications

Variants

Quit

Finish the work (function run()) and be ready to thread join

Event

Contains a portion of data for processing by the Worker T must has name and target

Busy

Inform Supervisor that worker become busy

Free

Inform Supervisor that worker become free for next event