Union lib::SsoString

#[repr(C)]
pub union SsoString {
    /* private fields */
}

Implementations

impl SsoString

pub fn new() -> Self

pub fn is_short(&self) -> bool

Returns true if this string is a short string (no heap allocations), and false otherwise

pub fn is_long(&self) -> bool

Returns !self.is_short()

pub fn tagged(&self) -> TaggedSsoString64<'_>

Returns the underlying union as an enum, allowing you to access the underlying short or long variant for the string

pub fn tagged_mut(&mut self) -> TaggedSsoString64Mut<'_>

Same as SsoString::tagged, but returns allows mutation of the underlying values instead

pub fn as_bytes(&self) -> &[u8]

Returns a slice of bytes of this string’s contents

pub fn as_mut_str(&mut self) -> &mut str

pub unsafe fn as_mut_vec(&mut self) -> &mut Vec<u8>

A method with the same name exists on std::string::String, but cannot exist on this version, as it does not use a vector internally (ever). I don’t think this method is good anyway, since it doesn’t scope the unsafe behaviour appropriately.

If you desperately want to use this kind of method, you can simply create &strs with unsafe unchecked methods. Probably not needed though, but I won’t judge :o)

pub fn as_str(&self) -> &str

pub fn capacity(&self) -> usize

pub fn clear(&mut self)

pub fn drain<R>(&mut self, _range: R) -> Drain<'_>
where R: RangeBounds<usize>,

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn extend_from_within<R>(&mut self, _src: R)
where R: RangeBounds<usize>,

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub unsafe fn from_raw_parts( buf: *mut u8, length: usize, capacity: usize ) -> Self

Creates a new SsoString::Long from a length, capacity and pointer. This method only exists to match std::string::String’s method of the same signature and name. It will always create a long string, which is probably what you want if you are using this method.

Safety (from std::string::String)

This is highly unsafe, due to the numer of invariants that aren’t checked:

The memory at buf needs to have been previously allocated by the same allocator the standard library uses, with a required alignment of exactly 1.

length needs to be less than or equal to capacity.

capacity needs to be the correct value.

The first length bytes at buf need to be valid UTF-8.

pub fn from_utf16(_v: &[u16]) -> Result<SsoString, FromUtf16Error>

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn from_utf16_lossy(_v: &[u16]) -> SsoString

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn from_utf8(_v: Vec<u8>) -> Result<String, FromUtf8Error>

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn from_utf8_lossy(_v: &[u8]) -> Cow<'_, SsoStr>

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub unsafe fn from_utf8_unchecked(_v: &[u8]) -> SsoString

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn insert(&mut self, _idx: usize, _c: char)

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn insert_str(&mut self, _idx: usize, _s: &str)

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn into_boxed_str(&self) -> Box<str>

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn leak<'a>(&self) -> &'a mut str

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn len(&self) -> usize

pub fn push(&mut self, ch: char)

pub fn push_str(&mut self, s: &str)

Push a str s onto the end of this string

pub fn remove(&mut self, _idx: usize) -> char

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn replace_range<R>(&mut self, _range: R, _replace_with: &str)
where R: RangeBounds<usize>,

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn reserve(&mut self, additional: usize)

pub unsafe fn set_len(&mut self, len: usize)

pub fn pop(&mut self) -> Option<char>

pub fn reserve_exact(&mut self, additional: usize)

This doesn’t actually reserve exactly additional extra bytes, it might allocate a few extra, just because of the implementation of Global.

pub fn retain<F>(&mut self, f: F)
where F: FnMut(char) -> bool,

Retains only the characters specified by the predicate.

TODO

This is a bad implementation of a simple function, it creates an entirely new string, it doesn’t require &mut. The better version would be easier to implement with as_mut_str.

I don’t think I’m going to bother with this any time soon.

pub fn with_capacity(capacity: usize) -> Self

pub fn shrink_to(&mut self, min_capacity: usize)

pub fn shrink_to_fit(&mut self)

This is currently just an alias for self.shrink_to(self.len()), it doesn’t avoid any branches just because it’s always valid

TODO

Implement this better

pub fn split_off(&mut self, _at: usize) -> String

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn truncate(&mut self, _new_len: usize)

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn try_reserve(&mut self, _additional: usize) -> Result<(), TryReserveError>

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call

pub fn try_reserve_exact( &mut self, _additional: usize ) -> Result<(), TryReserveError>

A method with the same name exists on std::string::String, but it is not yet implemented for SsoString. This function will panic on call