lib::vector
Trait Dot

pub trait Dot<T> {
    // Required method
    fn dot(&self, other: &Self) -> T;
}

A trait for any type that can have a dot product calculated from itself and another of Self.

Required Methods

fn dot(&self, other: &Self) -> T

Returns the dot product of self and other.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.