Crate lib

Modules

script

Confecciona um script padrão ou retorna de um arquivo externo indicado pela âncora .script.

style

Confecciona um CSS padrão ou retorna de um arquivo externo indicado pela âncora .css.

Structs

Presentation

Abstração de uma apresentação.

Enums

Handle

Possíveis handles que indicam como a próxima linha será processada.

Traits

Build

Constrói o arquivo final file.html como método do tipo Presentation. O nome do arquivo de entrada file será usado para nomear a saída.

Process

Opera sobre um tipo Presentation confeccionando o HTLM correspondente para cada String fornecida. O argumento line_no deve ser usado para identificar linhas onde ocorrem erros durante a confecção do HTML correspondente.

Functions

close_last_handle

Fecha o último handle para que a tag correspondente seja propriamente encerrada.

file_base64

Converte arquivo apontado para base64 e incorpora o raw data no HTML final.

generate_logo

Aponta uma logo para ser renderizada caso exista.

generate_mermaid_script

Fornece um script padrão para as âncoras .mermaid ou aceita um apontamento feito por .frommermaid

stv_to_html

Corta a extensão .stv e apensa .html o nome de arquivo que será usado como output.

table_generator

Processa linhas como rows de uma table em HTML.

trim_element

Remove a âncora da linha que a contém.

validate_stv

Valida o arquivo de entrada como um stv válido. Atualmente, apenas checka se o arquivo é um .stv.