Crate lib

Defines our shared libraries.

Protocol

Defines our can protocol and message enumerations for trivial parsing.

PID

Defines a simple PID controller.

Gain scheduled

Defines a PID controller that follows a parameter scheduled defined at compile time.

Wrapper

Defines some numerical wrappers for easier computations.

Modules

constants

Defines a simple set of constant identifiers for messages.

gain_scheduled

Defines a gain scheduling PID controller.

pid

Defines a PID controller.

protocol

Defines the protocol used to communicate over can bus.

wrapper

Defines a few wrapper types.

Macros

new_pid

Makes the instantiation of a new PID controller a bit more readable.