Crate lib

Structs

DataFrame

A dataframe struct similar to pythons pandas dataframe but with less functionality and only for integers The dataframe is stored as a vector of vectors The first vector contains the column names The rest of the vectors contain the data The data is stored as a vector of vectors

Equation

Equation struct that can be used to solve equations with two numbers and an operator (+, -, *, /, %, ^) Sol is the solution to the equation and is calculated when get_sol() is called.

EquationF

EquationF is an equation similar to Equation, but with f64 values. Sol is calculated when get_sol() is called.

ZeroEquation

ZeroEquation is an equation with a list of values on one side and a solution on the other. Similar to Equation, but with a list of values (in a Vec) that will add up to the solution (or be subtracted if the number is negative). Sol is set to 0 and values can be moved to the solution side of the equation with move_to_sol().

Functions

add

Returns the sum of two numbers.

add_float

Add for floating point numbers.

dataframe_create

Creates a new dataframe

divide

Returns the quotient of two numbers.

divide_float

Returns the quotient of two floating point numbers.

factorial

Returns the factorial of a number.

factorial_float

Returns the factorial of a floating point number.

fibonacci

Returns the fibonacci of a number.

fibonacci_float

Returns the fibbonacci of a floating point number.

is_even

Returns true if the number is even.

is_even_float

Returns true if the floating point number is even.

is_odd

Returns true if the number is odd.

is_odd_float

Returns true if the floating point number is odd.

is_perfect_cube

Returns true if the number is perfect cube.

is_perfect_cube_float

Returns true if the floating point number is perfect cube.

is_perfect_power

Returns true if the number is perfect power.

is_perfect_power_float

Returns true if the floating point number is perfect power.

is_perfect_square

Returns true if the number is positive.

is_perfect_square_float

Returns true if the floating point number is perfect square.

is_prime

Returns true if the number is prime.

is_prime_float

Returns true if the floating point number is prime.

modulo

Returns the remainder of two numbers.

modulo_float

Returns the remainder of two floating point numbers.

multiply

Multiply left and right numbers.

multiply_float

Multiply left and right floating point numbers.

power

Returns the left number raised to the right number.

power_float

Returns the left floating point number raised to the right floating point number.

root

Returns the left number to the root of the right number.

root_float

Returns the left floating point number to the root of the right floating point number.

subtract

Returns the difference of two numbers.

subtract_float

Returns the difference of two floating point numbers.

two_vector_median

Returns the median of two vectors

two_vector_median_float

Returns the median of two vectors float

two_vector_mode

Returns the mode of two vectors

two_vector_mode_float

Returns the mode of two vectors float.

vector_add

Returns the sum of two vectors

vector_add_float

vector_average

Returns the average of two vectors

vector_average_float

Returns the average of two vectors float

vector_divide

Returns the quotient of two vectors

vector_divide_float

Returns the quotient of two float vectors

vector_interquartile_range

Returns the inter quartile range of all elements in a vector

vector_interquartile_range_float

Returns the inter quartile range of all elements in a float vector

vector_max

Returns the maximum of two vectors

vector_max_float

Returns the maximum of two vectors float

vector_mean

Returns the mean of all elements in a vector

vector_mean_float

Returns the mean of all elements in a float vector

vector_median

Returns the median of all elements in a vector

vector_median_float

Returns the median of all elements in a float vector

vector_min

Returns the minimum of two vectors

vector_min_float

Returns the minimum of two vectors float

vector_mode

vector_modulo

Returns the remainder of two vectors

vector_modulo_float

Returns the modoulo of two vectors float

vector_multiply

Returns the product of two vectors

vector_multiply_float

Returns the product of two float vectors

vector_power

Returns the power of two vectors

vector_power_float

Returns the power of two vectors float

vector_product

Returns the product of all elements in a vector

vector_product_float

Returns the product of all elements in a float vector

vector_quartiles

Returns the quartiles of all elements in a vector

vector_quartiles_float

Returns the quartiles of all elements in a float vector

vector_range

Returns the range of all elements in a vector

vector_range_float

Returns the range of all elements in a float vector

vector_root

Returns the root of two vectors

vector_root_float

Returns the root of two vectors float

vector_standard_deviation

Returns the standard deviation of all elements in a vector

vector_standard_deviation_float

Returns the standard deviation of all elements in a float vector

vector_subtract

Returns the difference of two vectors

vector_subtract_float

Returns the difference of two float vectors

vector_sum

Returns the sum of all elements in a vector

vector_sum_float

Returns the sum of all elements in a float vector

vector_variance

Returns the variance of all elements in a vector

vector_variance_float

Returns the variance of all elements in a float vector