Crate lib

Structs

Image
Image struct returned from the read functions.

Info
Holds basic info about an image.

PngCustomChunk
PNG extension chunk.

Enums

ColFmt
Color format.

ColType
Color type – these are categories of color formats.

Functions

convert
Converts the image into a new allocation.

read
Reads an image and converts it to requested format.

read_bmp
Reads an image and converts it to requested format.

read_bmp_info
Returns width, height and color type of the image.

read_info
Returns width, height and color type of the image.

read_jpeg
Reads an image and converts it to requested format.

read_jpeg_info
Returns width, height and color type of the image.

read_png
Reads an image and converts it to requested format.

read_png_chunks
Like read_png but also returns the requested extension chunks.

read_png_info
Returns width, height and color type of the image.

read_tga
Reads an image and converts it to requested format.

read_tga_info
Returns width, height and color type of the image.

write
Writes an image and converts it to requested color type.

write_png
Writes an image and converts it to requested color type.

write_png_chunks
Like write_png but also writes the given extension chunks.

write_tga
Writes an image and converts it to requested color type.