Crate lib

Assertables: Rust crate of assert macros for testing

The assertables Rust crate provides many assert macros to improve your compile-time tests and run-time reliability.

Crate: https://crates.io/crates/assertables

Docs: https://docs.rs/assertables/

Repo: https://github.com/sixarm/assertables-rust-crate/

Contact: joel@joelparkerhenderson.com

Introduction

The Rust programming language provides a few built-in assert macros to test code:

assert!()
assert_eq!(a, b)
assert_ne!(a, b)

The assertables crate provides many more, so you can write smarter tests.

For values:

assert_gt!(a, b)
assert_lt!(a, b)

For numbers:

assert_in_delta!(a, b, delta)
assert_in_epsilon!(a, b, epsilon)

For strings:

assert_starts_with!(a, b)
assert_ends_with!(a, b)

For matching:

assert_contains!(a, b)
assert_is_match!(a, b)

For collections such as arrays, vectors, maps, sets:

assert_set_subset!(a, b)
assert_set_disjoint!(a, b)

For file system paths and input/output readers:

assert_fs_read_to_string_eq!(path1, path2)
assert_io_read_to_string_eq!(reader1, reader2)

For command capture of standard output and standard error:

assert_command_stdout_eq!(command1 stdout = command2 stdout);
assert_command_stderr_eq!(command1, command2);

Benefits

Your tests are more purposeful and powerful. This helps your code be more reliable.

Your assert failures provide more information. This helps you troubleshoot faster.

You gain runtime asserts. This helps you with validations and verifications.

Features

Easy to use: each macro is well-documented with runnable examples and tests.

Zero overhead: if you don’t use a macro, then it’s never compiled into your code.

Zero dependencies: the crate has no release dependencies, and just a short list of development dependencies.

Forms

Assertables macros come in three forms:

Panic macro: assert_* is for typical test uses with cargo test.

Debug macro: debug_assert_* is for runtime diagnostic configuration.

Result macro: assert_*_as_result is for runtime production configuration, such as for site reliability engineering, chaos engineering, validations, verifications, sanitizations, and more.

Naming conventions

Abbreviations:

eq ≈ equal

ne ≈ not equal.

lt ≈ less than

le ≈ less than or equal.

gt ≈ greater than

ge ≈ greater than or equal.

Shorthands:

path ≈ implements AsRef<Path> such as std::path::PathBuf.

reader ≈ implements method .read_to_string() such as std::io::Read.

matcher ≈ implements .is_match(…) such as regex::Regex.

containee ≈ usable inside .contains(…) such as a std::string::String substring.

set ≈ a collection such as ::std::collections::BTreeSet.

bag ≈ a collection such as ::std::collections::BTreeMap which has key counts.

Complete list of assert macros

assert_* for values

Compare values:

assert_eq!(a, b) ≈ a = b

assert_ne!(a, b) ≈ a ≠ b

assert_ge!(a, b) ≈ a ≥ b

assert_gt!(a, b) ≈ a > b

assert_le!(a, b) ≈ a ≤ b

assert_lt!(a, b) ≈ a < b

For infix operators

Compare values by using an infix value operator:

assert_infix!(a == b) ≈ a == b

assert_infix!(a != b) ≈ a ≠ b

assert_infix!(a < b) ≈ a < b

assert_infix!(a <= b) ≈ a ≤ b

assert_infix!(a > b) ≈ a > b

assert_infix!(a >= b) ≈ a ≥ b

Relate values by using an infix logical operator:

assert_infix!(a & b) ≈ a ∧ b ≈ a AND b

assert_infix!(a | b) ≈ a ∨ b ≈ a OR b

assert_infix!(a ^ b) ≈ a ⊻ b ≈ a XOR b

assert_infix!(a && b) ≈ a …∧ b ≈ a lazy AND b

assert_infix!(a || b) ≈ a …∨ b ≈ a lazy OR b

For nearness

Compare values by using nearness math conventions:

assert_in_delta!(a, b, delta) ≈ | a - b | ≤ delta

assert_in_epsilon(a, b, epsilon) ≈ | a - b | ≤ epsilon * min(a, b)

For strings

These macros help with strings and also other structures that provide matchers such as starts_with, ends_width, contains, and is_match. Each macro also has a corresponding not version.

assert_starts_with(a, b) ≈ a.starts_with(b)

assert_ends_with(a, b) ≈ a.ends_with(b)

assert_contains(container, containee) ≈ container.contains(containee)

assert_is_match(matcher, matchee) ≈ matcher.is_match(matchee)

For much more

There are many more macros that are conveniently grouped into modules.

For enums:

assert_option for Option (Some, None)

assert_result for Result (Ok, Err)

For collections, such as arrays, vectors, lists, maps:

assert_set for set collections

assert_bag for bag collections

For functions:

assert_fn for functions in general.

assert_fn_ok for functions that return Result::Ok.

assert_fn_err for functions that return Result::Err.

For readers:

assert_fs_read_to_string for file system path contents.

assert_io_read_to_string for input/output reader streams.

For external calls:

assert_command for commands and their stdout & stderr.

assert_program_args for programs with args and their stdout & stderr.

Forms

Forms for panic! versus Err()

The assert macros have three forms that you can use depending on your goals:

assert_gt!(a, b); // return () or panic!(…), for typical compile-time testing
 
debug_assert_gt!(a, b); // return () or panic!(…), for a non-optimized runtime
 
assert_gt_as_result!(a, b); // return Result Ok(()) or Err(…), for any runtime

Forms for messages

The assert macros have forms for default messages versus custom messages.

assert_gt!(1, 2); // panic!("assertion failed: assert_gt(1, 2)…")
 
assert_gt!(1, 2, "message"); // panic!("message")

Forms for comparing an other versus an expression

Some assert macros have forms for comparing an other versus an expression:

assert_io_read_to_string_eq!(reader1, reader2); // reader1.read_to_string() = reader2.read_to_string()
 
assert_io_read_to_string_eq_expr!(reader, expr); // reader1.read_to_string() = expr

Change highlights

Version 8

8.2:

Add assert_infix

Refactor into submodules for better discoverability and testability.

8.1:

Add Result macros assert_result_ok and assert_result_err

Add Option macros assert_option_some and assert_option_none

8.0:

Add assert_fs_read_to_string_* macros for comparing files.

Breaking change: rename assert_read_to_string_* macros to assert_io_read_to_string_*. If you use these macros, then please update your code to use the new naming convention.

Version 7

Add assert_in_delta, assert_in_epsilon.

Add assert_fn_* macros with multiple arities.

Add cargo release for optimized tagged releases.

Version 6

Add assert_starts_with, assert_ends_with, assert_contains, assert_is_match.

Add debug_assert_* macros everywhere.

Add GPL-3.0 license.

Tracking

Package: assertables-rust-crate

Version: 8.2.0

Created: 2021-03-30T15:47:49Z

Updated: 2024-09-04T20:21:53Z

License: MIT or Apache-2.0 or GPL-2.0 or GPL-3.0 or contact us for more

Contact: Joel Parker Henderson (joel@sixarm.com)

Modules

assert

Assert a condition is true.

assert_bag

Assert macros for comparing bags.

assert_command

Assert macros for comparing commands and their stdout & stderr.

assert_contains

Assert a container is a match for an expression.

assert_ends_with

Assert an expression (such as a string) ends with an expression (such as a string).

assert_eq

Assert an expression is equal to another.

assert_fn

Assert macros for comparing functions.

assert_fn_err

Assert macros for comparing functions that return errors.

assert_fn_ok

Assert macros for comparing functions that return Result::Ok.

assert_fs_read_to_string

Assert macros for comparing file system path contents.

assert_ge

Assert a value is greater than or equal to an expression.

assert_gt

Assert a value is greater than an expression.

assert_in_delta

Assert a number is within delta of another number.

assert_in_epsilon

Assert a number is within epsilon of another number.

assert_infix

Assert a infix operator, such as assert_infix!(a == b).

assert_io_read_to_string

Assert macros for comparing input/output reader streams.

assert_is_match

Assert a matcher is a match for an expression.

assert_le

Assert a value is less than or equal to an expression.

assert_lt

Assert a value is less than an expression.

assert_ne

Assert an expression is not equal to an expression.

assert_not_contains

Assert an expression (such as a string) does not contain an expression (such as a substring).

assert_not_ends_with

Assert an expression (such as a string) does not end with an expression (such as a string).

assert_not_match

Assert an expression (such as a regex) is not a match for an expression (such as a string).

assert_not_starts_with

Assert an expression (such as a string) does not start with an expression (such as a string).

assert_option

Assert macros for Option (Some or None)

assert_program_args

Assert macros for comparing programs with arguments.

assert_result

Assert macros for Result (Ok or Err)

assert_set

Assert macros for comparing set collections.

assert_starts_with

Assert an expression (such as a string) starts with an expression (such as a string).

Macros

assert_as_result

Assert a condition is true.

assert_bag_eq

Assert a bag is equal to another.

assert_bag_eq_as_result

Assert a bag is equal to another.

assert_bag_ne

Assert a bag is not equal to another.

assert_bag_ne_as_result

Assert a bag is not equal to another.

assert_bag_subbag

Assert a bag is a subbag of another.

assert_bag_subbag_as_result

Assert a bag is a subbag of another.

assert_bag_superbag

Assert a bag is a superbag of another.

assert_bag_superbag_as_result

Assert a bag is a superbag of another.

assert_command_stderr_contains

Assert a command stderr string contains a given containee.

assert_command_stderr_contains_as_result

Assert a command stderr string contains a given containee.

assert_command_stderr_eq

Assert a command stderr string is equal to another.

assert_command_stderr_eq_as_result

Assert a command stderr string is equal to another.

assert_command_stderr_eq_expr

Assert a command stderr string is equal to an expression.

assert_command_stderr_eq_expr_as_result

Assert a command stderr string is equal to an expression.

assert_command_stderr_is_match

Assert a command stderr string is a match to a regex.

assert_command_stderr_is_match_as_result

Assert a command stderr string is a match to a regex.

assert_command_stdout_contains

Assert a command stdout string contains a given containee.

assert_command_stdout_contains_as_result

Assert a command stdout string contains a given containee.

assert_command_stdout_eq

Assert a command stdout string is equal to another.

assert_command_stdout_eq_as_result

Assert a command stdout string is equal to another.

assert_command_stdout_eq_expr

Assert a command stdout string is equal to an expression.

assert_command_stdout_eq_expr_as_result

Assert a command stdout string is equal to an expression.

assert_command_stdout_is_match

Assert a command stdout string is a match to a regex.

assert_command_stdout_is_match_as_result

Assert a command stdout string is a match to a regex.

assert_contains

Assert a container is a match for an expression.

assert_contains_as_result

Assert an expression (such as a string) contains an expression (such as a substring).

assert_ends_with

Assert an expression (such as a string) ends with an expression (such as a string).

assert_ends_with_as_result

Assert an expression (such as a string) ends with an expression (such as a substring).

assert_eq_as_result

Assert an expression is equal to another.

assert_fn_eq

Assert a function output is equal to another function output.

assert_fn_eq_as_result

Assert a function output is equal to another function output.

assert_fn_eq_expr

Assert a function output is equal to an expression.

assert_fn_eq_expr_as_result

Assert a function output is equal to an expression.

assert_fn_err_eq

Assert a function err() is equal to another.

assert_fn_err_eq_as_result

Assert a function err() is equal to another.

assert_fn_err_eq_expr

Assert a function err() is equal to an expression.

assert_fn_err_eq_expr_as_result

Assert a function err() is equal to an expression.

assert_fn_err_ge

Assert a function err() is greater than or equal to another.

assert_fn_err_ge_as_result

Assert a function err() is greater than or equal to another.

assert_fn_err_ge_expr

Assert a function err() is greater than or equal to an expression.

assert_fn_err_ge_expr_as_result

Assert a function err() is greater than or equal to an expression.

assert_fn_err_gt

Assert a function err() is greater than another.

assert_fn_err_gt_as_result

Assert a function err() is greater than another.

assert_fn_err_gt_expr

Assert a function err() is greater than an expression.

assert_fn_err_gt_expr_as_result

Assert a function err() is greater than an expression.

assert_fn_err_le

Assert a function err() is less than or equal to another.

assert_fn_err_le_as_result

Assert a function err() is less than or equal to another.

assert_fn_err_le_expr

Assert a function err() is less than or equal to an expression.

assert_fn_err_le_expr_as_result

Assert a function err() is less than or equal to an expression.

assert_fn_err_lt

Assert a function err() is less than another.

assert_fn_err_lt_as_result

Assert a function err() is less than another.

assert_fn_err_lt_expr

Assert a function err() is less than an expression.

assert_fn_err_lt_expr_as_result

Assert a function err() is less than an expression.

assert_fn_err_ne

Assert a function err() is not equal to another.

assert_fn_err_ne_as_result

Assert a function err() is not equal to another.

assert_fn_err_ne_expr

Assert a function err() is not equal to an expression.

assert_fn_err_ne_expr_as_result

Assert a function err() is not equal to an expression.

assert_fn_ge

Assert a function output is greater than or equal to another.

assert_fn_ge_as_result

Assert a function output is greater than or equal to another.

assert_fn_ge_expr

Assert a function output is greater than or equal to an expression.

assert_fn_ge_expr_as_result

Assert a function output is greater than or equal to an expression.

assert_fn_gt

Assert a function output is greater than another.

assert_fn_gt_as_result

Assert a function output is greater than another.

assert_fn_gt_expr

Assert a function output is greater than an expression.

assert_fn_gt_expr_as_result

Assert a function output is greater than an expression.

assert_fn_le

Assert a function output is less than or equal to another.

assert_fn_le_as_result

Assert a function output is less than or equal to another.

assert_fn_le_expr

Assert a function output is less than or equal to an expression.

assert_fn_le_expr_as_result

Assert a function output is less than or equal to an expression.

assert_fn_lt

Assert a function output is less than another.

assert_fn_lt_as_result

Assert a function output is less than another.

assert_fn_lt_expr

Assert a function output is less than an expression.

assert_fn_lt_expr_as_result

Assert a function output is less than an expression.

assert_fn_ne

Assert a function output is not equal to another.

assert_fn_ne_as_result

Assert a function output is not equal to another.

assert_fn_ne_expr

Assert a function output is not equal to an expression.

assert_fn_ne_expr_as_result

Assert a function output is not equal to an expression.

assert_fn_ok_eq

Assert a function ok() is equal to another.

assert_fn_ok_eq_as_result

Assert a function ok() is equal to another.

assert_fn_ok_eq_expr

Assert a function ok() is equal to an expression.

assert_fn_ok_eq_expr_as_result

Assert a function ok() is equal to an expression.

assert_fn_ok_ge

Assert a function ok() is greater than or equal to another.

assert_fn_ok_ge_as_result

Assert a function ok() is greater than or equal to another.

assert_fn_ok_ge_expr

Assert a function ok() is greater than or equal to an expression.

assert_fn_ok_ge_expr_as_result

Assert a function ok() is greater than or equal to an expression.

assert_fn_ok_gt

Assert a function ok() is greater than another.

assert_fn_ok_gt_as_result

Assert a function ok() is greater than another.

assert_fn_ok_gt_expr

Assert a function ok() is greater than an expression.

assert_fn_ok_gt_expr_as_result

Assert a function ok() is greater than an expression.

assert_fn_ok_le

Assert a function ok() is less than or equal to another.

assert_fn_ok_le_as_result

Assert a function ok() is less than or equal to another.

assert_fn_ok_le_expr

Assert a function ok() is less than or equal to an expression.

assert_fn_ok_le_expr_as_result

Assert a function ok() is less than or equal to an expression.

assert_fn_ok_lt

Assert a function ok() is less than another.

assert_fn_ok_lt_as_result

Assert a function ok() is less than another.

assert_fn_ok_lt_expr

Assert a function ok() is less than an expression.

assert_fn_ok_lt_expr_as_result

Assert a function ok() is less than an expression.

assert_fn_ok_ne

Assert a function ok() is not equal to another.

assert_fn_ok_ne_as_result

Assert a function ok() is not equal to another.

assert_fn_ok_ne_expr

Assert a function ok() is not equal to an expression.

assert_fn_ok_ne_expr_as_result

Assert a function ok() is not equal to an expression.

assert_fs_read_to_string_contains

Assert a std::fs::read_to_string() contains a pattern.

assert_fs_read_to_string_contains_as_result

Assert a std::fs::read_to_string() contains a pattern.

assert_fs_read_to_string_eq

Assert a std::fs::read_to_string() value is equal to another.

assert_fs_read_to_string_eq_as_result

Assert a std::fs::read_to_string() is equal to another.

assert_fs_read_to_string_eq_expr

Assert a std::fs::read_to_string() value is equal to an expression.

assert_fs_read_to_string_eq_expr_as_result

Assert a std::fs::read_to_string() value is equal to an expression.

assert_fs_read_to_string_ge

Assert a std::fs::read_to_string() value is greater than or equal to another.

assert_fs_read_to_string_ge_as_result

Assert a std::fs::read_to_string() value is greater than or equal to another.

assert_fs_read_to_string_ge_expr

Assert a std::fs::read_to_string() value is greater than or equal to an expression.

assert_fs_read_to_string_ge_expr_as_result

Assert a std::fs::read_to_string() value is greater than or equal to an expression.

assert_fs_read_to_string_gt

Assert a std::fs::read_to_string() value is greater than another.

assert_fs_read_to_string_gt_as_result

Assert a std::fs::read_to_string() value is greater than another.

assert_fs_read_to_string_gt_expr

Assert a std::fs::read_to_string() value is greater than an expression.

assert_fs_read_to_string_gt_expr_as_result

Assert a std::fs::read_to_string() value is greater than an expression.

assert_fs_read_to_string_le

Assert a std::fs::read_to_string() value is less than or equal to another.

assert_fs_read_to_string_le_as_result

Assert a std::fs::read_to_string() value is less than or equal to another.

assert_fs_read_to_string_le_expr

Assert a std::fs::read_to_string() value is less than or equal to an expression.

assert_fs_read_to_string_le_expr_as_result

Assert a std::fs::read_to_string() value is less than or equal to an expression.

assert_fs_read_to_string_lt

Assert a std::fs::read_to_string() value is less than another.

assert_fs_read_to_string_lt_as_result

Assert a std::fs::read_to_string() value is less than another.

assert_fs_read_to_string_lt_expr

Assert a std::fs::read_to_string() value is less than an expression.

assert_fs_read_to_string_lt_expr_as_result

Assert a std::fs::read_to_string() value is less than an expression.

assert_fs_read_to_string_matches

Assert a std::fs::read_to_string() is a match to a regex.

assert_fs_read_to_string_matches_as_result

Assert a std::fs::read_to_string() is a match to a regex.

assert_fs_read_to_string_ne

Assert a std::fs::read_to_string() is not equal to another.

assert_fs_read_to_string_ne_as_result

Assert a std::fs::read_to_string() is not equal to another.

assert_fs_read_to_string_ne_expr

Assert a std::fs::read_to_string() is not equal to an expression.

assert_fs_read_to_string_ne_expr_as_result

Assert a std::fs::read_to_string() is not equal to an expression.

assert_ge

Assert a value is greater than or equal to an expression.

assert_ge_as_result

Assert a value is greater than or equal to an expression.

assert_gt

Assert a value is greater than an expression.

assert_gt_as_result

Assert a value is greater than an expression.

assert_in_delta

Assert a number is within delta of another number.

assert_in_delta_as_result

Assert a number is within delta of another number.

assert_in_epsilon

Assert a number is within epsilon of another number.

assert_in_epsilon_as_result

Assert a number is within epsilon of another number.

assert_infix

Assert a infix operator, such as assert_infix!(a == b).

assert_infix_as_result

Assert a infix operator, such as assert_infix!(a == b).

assert_io_read_to_string_contains

Assert a std::io::Read read_to_string() contains a pattern.

assert_io_read_to_string_contains_as_result

Assert a std::io::Read read_to_string() contains a pattern.

assert_io_read_to_string_eq

Assert a std::io::Read read_to_string() value is equal to another.

assert_io_read_to_string_eq_as_result

Assert a std::io::Read read_to_string() is equal to another.

assert_io_read_to_string_eq_expr

Assert a std::io::Read read_to_string() value is equal to an expression.

assert_io_read_to_string_eq_expr_as_result

Assert a std::io::Read read_to_string() value is equal to an expression.

assert_io_read_to_string_ge

Assert a std::io::Read read_to_string() value is greater than or equal to another.

assert_io_read_to_string_ge_as_result

Assert a std::io::Read read_to_string() value is greater than or equal to another.

assert_io_read_to_string_ge_expr

Assert a std::io::Read read_to_string() value is greater than or equal to an expression.

assert_io_read_to_string_ge_expr_as_result

Assert a std::io::Read read_to_string() value is greater than or equal to an expression.

assert_io_read_to_string_gt

Assert a std::io::Read read_to_string() value is greater than another.

assert_io_read_to_string_gt_as_result

Assert a std::io::Read read_to_string() value is greater than another.

assert_io_read_to_string_gt_expr

Assert a std::io::Read read_to_string() value is greater than an expression.

assert_io_read_to_string_gt_expr_as_result

Assert a std::io::Read read_to_string() value is greater than an expression.

assert_io_read_to_string_le

Assert a std::io::Read read_to_string() value is less than or equal to another.

assert_io_read_to_string_le_as_result

Assert a std::io::Read read_to_string() value is less than or equal to another.

assert_io_read_to_string_le_expr

Assert a std::io::Read read_to_string() value is less than or equal to an expression.

assert_io_read_to_string_le_expr_as_result

Assert a std::io::Read read_to_string() value is less than or equal to an expression.

assert_io_read_to_string_lt

Assert a std::io::Read read_to_string() value is less than another.

assert_io_read_to_string_lt_as_result

Assert a std::io::Read read_to_string() value is less than another.

assert_io_read_to_string_lt_expr

Assert a std::io::Read read_to_string() value is less than an expression.

assert_io_read_to_string_lt_expr_as_result

Assert a std::io::Read read_to_string() value is less than an expression.

assert_io_read_to_string_matches

Assert a std::io::Read read_to_string() is a match to a regex.

assert_io_read_to_string_matches_as_result

Assert a std::io::Read read_to_string() is a match to a regex.

assert_io_read_to_string_ne

Assert a std::io::Read read_to_string() is not equal to another.

assert_io_read_to_string_ne_as_result

Assert a std::io::Read read_to_string() is not equal to another.

assert_io_read_to_string_ne_expr

Assert a std::io::Read read_to_string() is not equal to an expression.

assert_io_read_to_string_ne_expr_as_result

Assert a std::io::Read read_to_string() is not equal to an expression.

assert_is_match

Assert a matcher is a match for an expression.

assert_is_match_as_result

Assert an expression (such as a regex) is a match for an expression (such as a string).

assert_le

Assert a value is less than or equal to an expression.

assert_le_as_result

Assert a value is less than or equal to an expression.

assert_lt

Assert a value is less than an expression.

assert_lt_as_result

Assert a value is less than an expression.

assert_ne_as_result

Assert an expression is not equal to an expression.

assert_not_contains

Assert an expression (such as a string) does not contain an expression (such as a substring).

assert_not_contains_as_result

Assert an expression (such as a string) does not contain an expression (such as a substring).

assert_not_ends_with

Assert an expression (such as a string) does not end with an expression (such as a string).

assert_not_ends_with_as_result

Assert an expression (such as a string) does not end with an expression (such as a substring).

assert_not_match

Assert an expression (such as a regex) is not a match for an expression (such as a string).

assert_not_match_as_result

Assert an expression (such as a regex) is not a match for an expression (such as a string).

assert_not_starts_with

Assert an expression (such as a string) does not start with an expression (such as a string).

assert_not_starts_with_as_result

Assert an expression (such as a string) does not start with an expression (such as a substring).

assert_option_none

Assert expression.is_none() is true.

assert_option_none_as_result

Assert an expression.is_none() is true.

assert_option_some

Assert expression.is_some() is true.

assert_option_some_as_result

Assert an expression.is_some() is true.

assert_program_args_stderr_contains

Assert a command (built with program and args) stderr string contains a given containee.

assert_program_args_stderr_contains_as_result

Assert a command (built with program and args) stderr string contains a given containee.

assert_program_args_stderr_eq

Assert a command (built with program and args) stderr string is equal to another.

assert_program_args_stderr_eq_as_result

Assert a command (built with program and args) stderr string is equal to another.

assert_program_args_stderr_eq_expr

Assert a command (built with program and args) stderr string is equal to an expression.

assert_program_args_stderr_eq_expr_as_result

Assert a command (built with program and args) stderr string is equal to an expression.

assert_program_args_stderr_ge

Assert a command (built with program and args) stderr string is greater than or equal to another.

assert_program_args_stderr_ge_as_result

Assert a command (built with program and args) stderr string is greater than or equal to another.

assert_program_args_stderr_ge_expr

Assert a command (built with program and args) stderr string is greater than or equal to an expression.

assert_program_args_stderr_ge_expr_as_result

Assert a command (built with program and args) stderr string is greater than or equal to an expression.

assert_program_args_stderr_gt

Assert a command (built with program and args) stderr string is greater than to another.

assert_program_args_stderr_gt_as_result

Assert a command (built with program and args) stderr string is greater than another.

assert_program_args_stderr_gt_expr

Assert a command (built with program and args) stderr string is greater than an expression.

assert_program_args_stderr_gt_expr_as_result

Assert a command (built with program and args) stderr string is greater than an expression.

assert_program_args_stderr_is_match

Assert a command (built with program and args) stderr string is a match to a regex.

assert_program_args_stderr_is_match_as_result

Assert a command (built with program and args) stderr string is a match to a regex.

assert_program_args_stderr_le

Assert a command (built with program and args) stderr string is less than or equal to another.

assert_program_args_stderr_le_as_result

Assert a command (built with program and args) stderr string is less than or equal to another.

assert_program_args_stderr_le_expr

Assert a command (built with program and args) stderr string is less than or equal to an expression.

assert_program_args_stderr_le_expr_as_result

Assert a command (built with program and args) stderr string is less than or equal to an expression.

assert_program_args_stderr_lt

Assert a command (built with program and args) stderr string is less than another.

assert_program_args_stderr_lt_as_result

Assert a command (built with program and args) stderr string is less than another.

assert_program_args_stderr_lt_expr

Assert a command (built with program and args) stderr string is less than an expression.

assert_program_args_stderr_lt_expr_as_result

Assert a command (built with program and args) stderr string is less than an expression.

assert_program_args_stderr_ne

Assert a command (built with program and args) stderr string is not equal to another.

assert_program_args_stderr_ne_as_result

Assert a command (built with program and args) stderr string is not equal to another.

assert_program_args_stderr_ne_expr

Assert a command (built with program and args) stderr string is not equal to an expression.

assert_program_args_stderr_ne_expr_as_result

Assert a command (built with program and args) stderr string is not equal to an expression.

assert_program_args_stdout_contains

Assert a command (built with program and args) stdout string contains a given containee.

assert_program_args_stdout_contains_as_result

Assert a command (built with program and args) stdout string contains a given containee.

assert_program_args_stdout_eq

Assert a command (built with program and args) stdout string is equal to another.

assert_program_args_stdout_eq_as_result

Assert a command (built with program and args) stdout string is equal to another.

assert_program_args_stdout_eq_expr

Assert a command (built with program and args) stdout string is equal to an expression.

assert_program_args_stdout_eq_expr_as_result

Assert a command (built with program and args) stdout string is equal to an expression.

assert_program_args_stdout_ge

Assert a command (built with program and args) stdout string is greater than or equal to another.

assert_program_args_stdout_ge_as_result

Assert a command (built with program and args) stdout string is greater than or equal to another.

assert_program_args_stdout_ge_expr

Assert a command (built with program and args) stdout string is greater than or equal to an expression.

assert_program_args_stdout_ge_expr_as_result

Assert a command (built with program and args) stdout string is greater than or equal to an expression.

assert_program_args_stdout_gt

Assert a command (built with program and args) stdout string is greater than another.

assert_program_args_stdout_gt_as_result

Assert a command (built with program and args) stdout string is greater than to another.

assert_program_args_stdout_gt_expr

Assert a command (built with program and args) stdout string is greater than an expression.

assert_program_args_stdout_gt_expr_as_result

Assert a command (built with program and args) stdout string is greater than an expression.

assert_program_args_stdout_is_match

Assert a command (built with program and args) stdout string is a match to a regex.

assert_program_args_stdout_is_match_as_result

Assert a command (built with program and args) stdout string is a match to a regex.

assert_program_args_stdout_le

Assert a command (built with program and args) stdout string is less than or equal to another.

assert_program_args_stdout_le_as_result

Assert a command (built with program and args) stdout string is less than or equal to another.

assert_program_args_stdout_le_expr

Assert a command (built with program and args) stdout string is less than or equal to an expression.

assert_program_args_stdout_le_expr_as_result

Assert a command (built with program and args) stdout string is less than or equal to an expression.

assert_program_args_stdout_lt

Assert a command (built with program and args) stdout string is less than another.

assert_program_args_stdout_lt_as_result

Assert a command (built with program and args) stdout string is less than another.

assert_program_args_stdout_lt_expr

Assert a command (built with program and args) stdout string is less than an expression.

assert_program_args_stdout_lt_expr_as_result

Assert a command (built with program and args) stdout string is less than an expression.

assert_program_args_stdout_ne

Assert a command (built with program and args) stdout string is not equal to another.

assert_program_args_stdout_ne_as_result

Assert a command (built with program and args) stdout string is not equal to another.

assert_program_args_stdout_ne_expr

Assert a command (built with program and args) stdout string is not equal to an expression.

assert_program_args_stdout_ne_expr_as_result

Assert a command (built with program and args) stdout string is not equal to an expression.

assert_result_err

Assert expression.is_err() is true.

assert_result_err_as_result

Assert an expression.is_err() is true.

assert_result_ok

Assert expression.is_ok() is true.

assert_result_ok_as_result

Assert an expression.is_ok() is true.

assert_set_disjoint

Assert a set is disjoint with another.

assert_set_disjoint_as_result

Assert a set is disjoint with another.

assert_set_eq

Assert a set is equal to another.

assert_set_eq_as_result

Assert a set is equal to another.

assert_set_joint

Assert a set is joint with another.

assert_set_joint_as_result

Assert a set is joint with another.

assert_set_ne

Assert a set is not equal to another.

assert_set_ne_as_result

Assert a set is not equal to another.

assert_set_subset

Assert a set is a subset of another.

assert_set_subset_as_result

Assert a set is a subset of another.

assert_set_superset

Assert a set is a superset of another.

assert_set_superset_as_result

Assert a set is a superset of another.

assert_starts_with

Assert an expression (such as a string) starts with an expression (such as a string).

assert_starts_with_as_result

Assert an expression (such as a string) starts with an expression (such as a substring).

debug_assert_bag_eq

Assert a bag is equal to another.

debug_assert_bag_ne

Assert a bag is not equal to another.

debug_assert_bag_subbag

Assert a bag is a subbag of another.

debug_assert_bag_superbag

Assert a bag is a superbag of another.

debug_assert_command_stderr_contains

Assert a command stderr string contains a given containee.

debug_assert_command_stderr_eq

Assert a command stderr string is equal to another.

debug_assert_command_stderr_eq_expr

Assert a command stderr string is equal to an expression.

debug_assert_command_stderr_is_match

Assert a command stderr string is a match to a regex.

debug_assert_command_stdout_contains

Assert a command stdout string contains a given containee.

debug_assert_command_stdout_eq

Assert a command stdout string is equal to another.

debug_assert_command_stdout_eq_expr

Assert a command stdout string is equal to an expression.

debug_assert_command_stdout_is_match

Assert a command stdout string is a match to a regex.

debug_assert_contains

Assert a container is a match for an expression.

debug_assert_ends_with

Assert an expression (such as a string) ends with an expression (such as a string).

debug_assert_fn_eq

Assert a function output is equal to another function output.

debug_assert_fn_eq_expr

Assert a function output is equal to an expression.

debug_assert_fn_err_eq

Assert a function err() is equal to another.

debug_assert_fn_err_eq_expr

Assert a function err() is equal to an expression.

debug_assert_fn_err_ge

Assert a function err() is greater than or equal to another.

debug_assert_fn_err_ge_expr

Assert a function err() is greater than or equal to an expression.

debug_assert_fn_err_gt

Assert a function err() is greater than another.

debug_assert_fn_err_gt_expr

Assert a function err() is greater than an expression.

debug_assert_fn_err_le

Assert a function err() is less than or equal to another.

debug_assert_fn_err_le_expr

Assert a function err() is less than or equal to an expression.

debug_assert_fn_err_lt

Assert a function err() is less than another.

debug_assert_fn_err_lt_expr

Assert a function err() is less than an expression.

debug_assert_fn_err_ne

Assert a function err() is not equal to another.

debug_assert_fn_err_ne_expr

Assert a function err() is not equal to an expression.

debug_assert_fn_ge

Assert a function output is greater than or equal to another.

debug_assert_fn_ge_expr

Assert a function output is greater than or equal to an expression.

debug_assert_fn_gt

Assert a function output is greater than another.

debug_assert_fn_gt_expr

Assert a function output is greater than an expression.

debug_assert_fn_le

Assert a function output is less than or equal to another.

debug_assert_fn_le_expr

Assert a function output is less than or equal to an expression.

debug_assert_fn_lt

Assert a function output is less than another.

debug_assert_fn_lt_expr

Assert a function output is less than an expression.

debug_assert_fn_ne

Assert a function output is not equal to another.

debug_assert_fn_ne_expr

Assert a function output is not equal to an expression.

debug_assert_fn_ok_eq

Assert a function ok() is equal to another.

debug_assert_fn_ok_eq_expr

Assert a function ok() is equal to an expression.

debug_assert_fn_ok_ge

Assert a function ok() is greater than or equal to another.

debug_assert_fn_ok_ge_expr

Assert a function ok() is greater than or equal to an expression.

debug_assert_fn_ok_gt

Assert a function ok() is greater than another.

debug_assert_fn_ok_gt_expr

Assert a function ok() is greater than an expression.

debug_assert_fn_ok_le

Assert a function ok() is less than or equal to another.

debug_assert_fn_ok_le_expr

Assert a function ok() is less than or equal to an expression.

debug_assert_fn_ok_lt

Assert a function ok() is less than another.

debug_assert_fn_ok_lt_expr

Assert a function ok() is less than an expression.

debug_assert_fn_ok_ne

Assert a function ok() is not equal to another.

debug_assert_fn_ok_ne_expr

Assert a function ok() is not equal to an expression.

debug_assert_fs_read_to_string_contains

Assert a std::fs::read_to_string() contains a pattern.

debug_assert_fs_read_to_string_eq

Assert a std::fs::read_to_string() value is equal to another.

debug_assert_fs_read_to_string_eq_expr

Assert a std::fs::read_to_string() value is equal to an expression.

debug_assert_fs_read_to_string_ge

Assert a std::fs::read_to_string() value is greater than or equal to another.

debug_assert_fs_read_to_string_ge_expr

Assert zzz.

debug_assert_fs_read_to_string_gt

Assert a std::fs::read_to_string() value is greater than another.

debug_assert_fs_read_to_string_gt_expr

Assert a std::fs::read_to_string() value is greater than an expression.

debug_assert_fs_read_to_string_le

Assert a std::fs::read_to_string() value is less than or equal to another.

debug_assert_fs_read_to_string_le_expr

Assert a std::fs::read_to_string() value is less than or equal to an expression.

debug_assert_fs_read_to_string_lt

Assert a std::fs::read_to_string() value is less than another.

debug_assert_fs_read_to_string_lt_expr

Assert a std::fs::read_to_string() value is less than an expression.

debug_assert_fs_read_to_string_matches

Assert a std::fs::read_to_string() is a match to a regex.

debug_assert_fs_read_to_string_ne

Assert a std::fs::read_to_string() is not equal to another.

debug_assert_fs_read_to_string_ne_expr

Assert a std::fs::read_to_string() is not equal to an expression.

debug_assert_ge

Assert a value is greater than or equal to an expression.

debug_assert_gt

Assert a value is greater than an expression.

debug_assert_in_delta

Assert a number is within delta of another number.

debug_assert_in_epsilon

Assert a number is within epsilon of another number.

debug_assert_infix

Assert a infix operator, such as assert_infix!(a == b).

debug_assert_io_read_to_string_contains

Assert a std::io::Read read_to_string() contains a pattern.

debug_assert_io_read_to_string_eq

Assert a std::io::Read read_to_string() value is equal to another.

debug_assert_io_read_to_string_eq_expr

Assert a std::io::Read read_to_string() value is equal to an expression.

debug_assert_io_read_to_string_ge

Assert a std::io::Read read_to_string() value is greater than or equal to another.

debug_assert_io_read_to_string_ge_expr

Assert zzz.

debug_assert_io_read_to_string_gt

Assert a std::io::Read read_to_string() value is greater than another.

debug_assert_io_read_to_string_gt_expr

Assert a std::io::Read read_to_string() value is greater than an expression.

debug_assert_io_read_to_string_le

Assert a std::io::Read read_to_string() value is less than or equal to another.

debug_assert_io_read_to_string_le_expr

Assert a std::io::Read read_to_string() value is less than or equal to an expression.

debug_assert_io_read_to_string_lt

Assert a std::io::Read read_to_string() value is less than another.

debug_assert_io_read_to_string_lt_expr

Assert a std::io::Read read_to_string() value is less than an expression.

debug_assert_io_read_to_string_matches

Assert a std::io::Read read_to_string() is a match to a regex.

debug_assert_io_read_to_string_ne

Assert a std::io::Read read_to_string() is not equal to another.

debug_assert_io_read_to_string_ne_expr

Assert a std::io::Read read_to_string() is not equal to an expression.

debug_assert_is_match

Assert a matcher is a match for an expression.

debug_assert_le

Assert a value is less than or equal to an expression.

debug_assert_lt

Assert a value is less than an expression.

debug_assert_not_contains

Assert an expression (such as a string) does not contain an expression (such as a substring).

debug_assert_not_ends_with

Assert an expression (such as a string) does not end with an expression (such as a string).

debug_assert_not_match

Assert an expression (such as a regex) is not a match for an expression (such as a string).

debug_assert_not_starts_with

Assert an expression (such as a string) does not start with an expression (such as a string).

debug_assert_option_none

Assert expression.is_none() is true.

debug_assert_option_some

Assert expression.is_some() is true.

debug_assert_program_args_stderr_contains

Assert a command (built with program and args) stderr string contains a given containee.

debug_assert_program_args_stderr_eq

Assert a command (built with program and args) stderr string is equal to another.

debug_assert_program_args_stderr_eq_expr

Assert a command (built with program and args) stderr string is equal to an expression.

debug_assert_program_args_stderr_ge

Assert a command (built with program and args) stderr string greater than or equal to another.

debug_assert_program_args_stderr_ge_expr

Assert a command (built with program and args) stderr string is greater than or equal to an expression.

debug_assert_program_args_stderr_gt

Assert a command (built with program and args) stderr string is equal to another.

debug_assert_program_args_stderr_gt_expr

Assert a command (built with program and args) stderr string is greater than an expression.

debug_assert_program_args_stderr_is_match

Assert a command (built with program and args) stderr string is a match to a regex.

debug_assert_program_args_stderr_le

Assert a command (built with program and args) stderr string is less than or equal to another.

debug_assert_program_args_stderr_le_expr

Assert a command (built with program and args) stderr string is equal to an expression.

debug_assert_program_args_stderr_lt

Assert a command (built with program and args) stderr string is less than another.

debug_assert_program_args_stderr_lt_expr

Assert a command (built with program and args) stderr string is less than an expression.

debug_assert_program_args_stderr_ne

Assert a command (built with program and args) stderr string is not equal to another.

debug_assert_program_args_stderr_ne_expr

Assert a command (built with program and args) stderr string is not equal to an expression.

debug_assert_program_args_stdout_contains

Assert a command (built with program and args) stdout string contains a given containee.

debug_assert_program_args_stdout_eq

Assert a command (built with program and args) stdout string is equal to another.

debug_assert_program_args_stdout_eq_expr

Assert a command (built with program and args) stdout string is equal to an expression.

debug_assert_program_args_stdout_ge

Assert a command (built with program and args) stdout string is greater than or equal to another.

debug_assert_program_args_stdout_ge_expr

Assert a command (built with program and args) stdout string is greater than or equal to an expression.

debug_assert_program_args_stdout_gt

Assert a command (built with program and args) stdout string is greater than another.

debug_assert_program_args_stdout_gt_expr

Assert a command (built with program and args) stdout string is greater than an expression.

debug_assert_program_args_stdout_is_match

Assert a command (built with program and args) stdout string is a match to a regex.

debug_assert_program_args_stdout_le

Assert a command (built with program and args) stdout string is less than or equal to another.

debug_assert_program_args_stdout_le_expr

Assert a command (built with program and args) stdout string is less than or equal to an expression.

debug_assert_program_args_stdout_lt

Assert a command (built with program and args) stdout string is less than another.

debug_assert_program_args_stdout_lt_expr

Assert a command (built with program and args) stdout string is less than an expression.

debug_assert_program_args_stdout_ne

Assert a command (built with program and args) stdout string is not equal to another.

debug_assert_program_args_stdout_ne_expr

Assert a command (built with program and args) stdout string is not equal to an expression.

debug_assert_result_err

Assert expression.is_err() is true.

debug_assert_result_ok

Assert expression.is_ok() is true.

debug_assert_set_disjoint

Assert a set is disjoint with another.

debug_assert_set_eq

Assert a set is equal to another.

debug_assert_set_joint

Assert a set is joint with another.

debug_assert_set_ne

Assert a set is not equal to another.

debug_assert_set_subset

Assert a set is a subset of another.

debug_assert_set_superset

Assert a set is a superset of another.

debug_assert_starts_with

Assert an expression (such as a string) starts with an expression (such as a string).