Crate lib

Assertables: Rust crate of assert macros for testing

The assertables Rust crate provides many assert macros to improve your compile-time tests and run-time reliability.

Crate: https://crates.io/crates/assertables

Docs: https://docs.rs/assertables/

Repo: https://github.com/sixarm/assertables-rust-crate/

Why use this?

When you write Rust tests, then you can use Rust assert macros, such as:

assert_eq!(value1, value2)

The assertables Rust crate provides many more assert macros for values, strings, vectors, paths, readers, commands, and more, such as:

// value1 greater than value2
assert_gt!(value1, value2);

// string1 starts with string2
assert_starts_with!(string1, string2);

// regex is match of string
assert_is_match!(regex, string);

// vector1 as set ⊆ vector2 as set
assert_set_subset!(vector1, vector2);

// function1 ok = function2 ok
assert_fn_ok_eq!(function1, function2);

// path1 as string = path2 as string
assert_fs_read_to_string_eq!(path1, path2);

// reader1 as string = reader2 as string
assert_io_read_to_string_eq!(reader1, reader2);

// command1 standard output = command2 standard output
assert_command_stdout_eq!(command1, command2);

See below for the complete list of all the assert macros.

Benefits

Your tests are more purposeful and powerful, which helps your code be more reliable.

Your assert failures provide more information, which helps you troubleshoot faster.

You gain runtime asserts, which helps you with validations and verifications.

Features

Easy to use: each macro is well-documented with runnable examples and tests.

Zero overhead: if you don’t use a macro, then it’s never compiled into your code.

Three forms: assert_* for development, debug_assert_* for debugging, and assert_*_as_result for production.

Complete list of assert macros

assert_* for values

Compare values:

assert_eq!(a, b) ≈ a = b

assert_ne!(a, b) ≈ a ≠ b

assert_ge!(a, b) ≈ a ≥ b

assert_gt!(a, b) ≈ a > b

assert_le!(a, b) ≈ a ≤ b

assert_lt!(a, b) ≈ a < b

Compare values by using nearness:

assert_in_delta!(a, b, delta) ≈ | a - b | ≤ delta

assert_in_epsilon(a, b, epsilon) ≈ | a - b | ≤ epsilon * min(a, b)

assert_* for strings and matchers

These macros help with strings and also other structures that provide matchers such as starts_with, ends_width, contains, and is_match.

assert_starts_with(a, b) ≈ a.starts_with(b)

assert_not_starts_with(a, b) ≈ !a.starts_with(b)

assert_ends_with(a, b) ≈ a.ends_with(b)

assert_not_ends_with(a, b) ≈ !a.ends_with(b)

assert_contains(container, containee) ≈ container.contains(containee)

assert_not_contains(container, containee) ≈ !container.contains(containee)

assert_is_match(matcher, matchee) ≈ matcher.is_match(matchee)

assert_not_match(matcher, matchee) ≈ !matcher.is_match(matchee)

assert_set_* for set collection comparisons

These macros help with comparison of set parameters, such as two arrays or two vectors. where the item order does not matter, and the item count does not matter. These macros convert their inputs into HashSet iterators.

assert_set_eq!(a, b) ≈ set a = set b

assert_set_ne!(a, b) ≈ set a ≠ set b

assert_set_subset!(a, b) ≈ set a ⊆ set b

assert_set_superset!(a, b) ≈ set a ⊇ set b

assert_set_joint!(a, b) ≈ set a ∩ set b ≠ ∅

assert_set_disjoint!(a, b) ≈ set a ∩ set b = ∅

assert_bag_* for bag collection comparisons

These macros help with comparison of bag parameters, such as comparison of two arrays or two vectors, where the item order does not matter, and the item count does matter. These macros convert their inputs into HashMap iterators.

assert_bag_eq(a, b) ≈ bag a = bag b

assert_bag_ne(a, b) ≈ bag a ≠ bag b

assert_bag_subbag(a, b) ≈ bag a ⊆ bag b

assert_bag_superbag(a, b) ≈ bag a ⊇ bag b

assert_fn_* for function return-value comparisons

Compare a function with another function:

assert_fn_eq!(function1, function2) ≈ function1() = function2()

assert_fn_ne!(function1, function2) ≈ function1() ≠ function2()

assert_fn_ge!(function1, function2) ≈ function1() ≥ function2()

assert_fn_gt!(function1, function2) ≈ function1() > function2()

assert_fn_le!(function1, function2) ≈ function1() ≤ function2()

assert_fn_lt!(function1, function2) ≈ function1() < function2()

Compare a function with an expression:

assert_fn_eq_expr!(function, expr) ≈ function() = expr

assert_fn_ne_expr!(function, expr) ≈ function() ≠ expr

assert_fn_ge_expr!(function, expr) ≈ function() ≥ expr

assert_fn_gt_expr!(function, expr) ≈ function() > expr

assert_fn_le_expr!(function, expr) ≈ function() ≤ expr

assert_fn_lt_expr!(function, expr) ≈ function() < expr

assert_fn_ok_* for Result Ok() comparisons

Compare a function Ok() with another function Ok():

assert_fn_ok_eq!(function1, function2) ≈ function1().ok().unwrap() = function2().ok().unwrap()

assert_fn_ok_ne!(function1, function2) ≈ function1().ok().unwrap() ≠ function2().ok().unwrap()

assert_fn_ok_ge!(function1, function2) ≈ function1().ok().unwrap() ≥ function2().ok().unwrap()

assert_fn_ok_gt!(function1, function2) ≈ function1().ok().unwrap() > function2().ok().unwrap()

assert_fn_ok_le!(function1, function2) ≈ function1().ok().unwrap() ≤ function2().ok().unwrap()

assert_fn_ok_lt!(function1, function2) ≈ function1().ok().unwrap() < function2().ok().unwrap()

Compare a function Ok() with an expression:

assert_fn_ok_eq_expr!(function, expr) ≈ function().ok().unwrap() = expr

assert_fn_ok_ne_expr!(function, expr) ≈ function().ok().unwrap() ≠ expr

assert_fn_ok_ge_expr!(function, expr) ≈ function().ok().unwrap() ≥ expr

assert_fn_ok_gt_expr!(function, expr) ≈ function().ok().unwrap() > expr

assert_fn_ok_le_expr!(function, expr) ≈ function().ok().unwrap() ≤ expr

assert_fn_ok_lt_expr!(function, expr) ≈ function().ok().unwrap() < expr

assert_fn_err_* for function Err() comparisons

Compare a function Err() with another function Err():

assert_fn_err_eq!(function1, function2) ≈ function1().unwrap_err() = function2().unwrap_err()

assert_fn_err_ne!(function1, function2) ≈ function1().unwrap_err() ≠ function2().unwrap_err()

assert_fn_err_ge!(function1, function2) ≈ function1().unwrap_err() ≥ function2().unwrap_err()

assert_fn_err_gt!(function1, function2) ≈ function1().unwrap_err() > function2().unwrap_err()

assert_fn_err_le!(function1, function2) ≈ function1().unwrap_err() ≤ function2().unwrap_err()

assert_fn_err_lt!(function1, function2) ≈ function1().unwrap_err() < function2().unwrap_err()

Compare a function Err() with an expression:

assert_fn_err_eq!(function, expr) ≈ function().unwrap_err() = expr

assert_fn_err_ne!(function, expr) ≈ function().unwrap_err() ≠ expr

assert_fn_err_ge!(function, expr) ≈ function().unwrap_err() ≥ expr

assert_fn_err_gt!(function, expr) ≈ function().unwrap_err() > expr

assert_fn_err_le!(function, expr) ≈ function().unwrap_err() ≤ expr

assert_fn_err_lt!(function, expr) ≈ function().unwrap_err() < expr

assert_fs_read_to_string_* for std::fs path comparisons

These macros help with file system paths, such as disk files, Path, PathBuf, the trait AsRef<Path>, and anything that is readable via std::fs::read_to_string(…).

Compare a path with another path:

assert_fs_read_to_string_eq!(path1, path2) ≈ std::fs::read_to_string(path1) = std::fs::read_to_string(path2)

assert_fs_read_to_string_ne!(path1, path2) ≈ std::fs::read_to_string(path1) ≠ std::fs::read_to_string(path2)

assert_fs_read_to_string_ge!(path1, path2) ≈ std::fs::read_to_string(path1) ≥ std::fs::read_to_string(path2)

assert_fs_read_to_string_gt!(path1, path2) ≈ std::fs::read_to_string(path1) > std::fs::read_to_string(path2)

assert_fs_read_to_string_le!(path1, path2) ≈ std::fs::read_to_string(path1) ≤ std::fs::read_to_string(path2)

assert_fs_read_to_string_lt!(path1, path2) ≈ std::fs::read_to_string(path1) < std::fs::read_to_string(path2)

Compare a path with an expression:

assert_fs_read_to_string_eq_expr(path, expr) ≈ std::fs::read_to_string(path) = expr

assert_fs_read_to_string_ne_expr(path, expr) ≈ std::fs::read_to_string(path) ≠ expr

assert_fs_read_to_string_ge_expr(path, expr) ≈ std::fs::read_to_string(path) ≥ expr

assert_fs_read_to_string_gt_expr(path, expr) ≈ std::fs::read_to_string(path) > expr

assert_fs_read_to_string_le_expr(path, expr) ≈ std::fs::read_to_string(path) ≤ expr

assert_fs_read_to_string_lt_expr(path, expr) ≈ std::fs::read_to_string(path) < expr

Compare a path with its contents:

assert_fs_read_to_string_contains(path, containee) ≈ std::fs::read_to_string(path).contains(containee)

assert_fs_read_to_string_matches(path, matcher) ≈ matcher.is_match(std::fs::read_to_string(path))

assert_io_read_to_string_* for std::io::Read comparisons

These macros help with readers, such as file handles, byte arrays, input streams, and the trait std::io::Read.

Compare a reader with another reader:

assert_io_read_to_string_eq!(reader1, reader2) ≈ reader1.read_to_string() = reader2.read_to_string()

assert_io_read_to_string_ne!(reader1, reader2) ≈ reader1.read_to_string() ≠ reader2.read_to_string()

assert_io_read_to_string_ge!(reader1, reader2) ≈ reader1.read_to_string() ≥ reader2.read_to_string()

assert_io_read_to_string_gt!(reader1, reader2) ≈ reader1.read_to_string() > reader2.read_to_string()

assert_io_read_to_string_le!(reader1, reader2) ≈ reader1.read_to_string() ≤ reader2.read_to_string()

assert_io_read_to_string_lt!(reader1, reader2) ≈ reader1.read_to_string() < reader2.read_to_string()

Compare a reader with an expression:

assert_io_read_to_string_eq_expr(reader, expr) ≈ reader.read_to_string() = expr

assert_io_read_to_string_ne_expr(reader, expr) ≈ reader.read_to_string() ≠ expr

assert_io_read_to_string_ge_expr(reader, expr) ≈ reader.read_to_string() ≥ expr

assert_io_read_to_string_gt_expr(reader, expr) ≈ reader.read_to_string() > expr

assert_io_read_to_string_le_expr(reader, expr) ≈ reader.read_to_string() ≤ expr

assert_io_read_to_string_lt_expr(reader, expr) ≈ reader.read_to_string() < expr

Compare a reader with its contents:

assert_io_read_to_string_contains(reader, containee) ≈ reader.read_to_string().contains(containee)

assert_io_read_to_string_matches(reader, matcher) ≈ matcher.is_match(reader.read_to_string())

assert_command_* for process command comparisons

Compare command standard output string:

assert_command_stdout_eq!(command1, command2) ≈ command1 stdout = command2 stdout

assert_command_stdout_eq_expr!(command, expr) ≈ command stdout = expr

assert_command_stdout_contains!(command, containee) ≈ command stdout contains containee

assert_command_stdout_is_match!(command, matcher) ≈ command stdout is a matcher match

Compare command standard error string:

assert_command_stderr_eq!(command1, command2) ≈ command1 stderr = command2 stderr

assert_command_stderr_eq_expr!(command, expr) ≈ command stderr = expr

assert_command_stderr_contains!(command, containee) ≈ command stderr contains containee

assert_command_stderr_is_match!(command, matcher) ≈ command stderr is a matcher match

assert_program_args_* for process command comparisons created via program name and args interator

Compare command using program and arguments to standard output:

assert_program_args_stdout_eq!(program1, args1, program2, args2) ≈ command using program1 and args1 to stdout = command2 with program2 and args2 to stdout

assert_program_args_stdout_eq_expr!(program, args, expr) ≈ command using program and args to stdout = expr

assert_program_args_stdout_contains!(program, args, containee) ≈ command using program and args to stdout contains containee

assert_program_args_stdout_is_match!(program, args, matcher) ≈ matcher is match with command using program and args

Compare command using program and arguments to standard output:

assert_program_args_stderr_eq!(program1, args1, program2, args2) ≈ command using program1 and args1 to stderr = command2 with program2 and args2 to stderr

assert_program_args_stderr_eq_expr!(program, args, expr) ≈ command using program and args to stderr = expr

assert_program_args_stderr_contains!(program, args, containee) ≈ command using program and args to stderr contains containee

assert_program_args_stderr_is_match!(program, args, matcher) ≈ matcher is match with command using program and args

Naming conventions

Abbreviations:

eq ≈ equal

ne ≈ not equal.

ge ≈ greater than or equal.

gt ≈ greater than

le ≈ less than or equal.

lt ≈ less than

Shorthands:

path ≈ implements AsRef<Path> such as std::path::PathBuf.

reader ≈ implements method .read_to_string() such as std::io::Read.

matcher ≈ implements .is_match(…) such as regex::Regex.

containee ≈ usable inside .contains(…) such as a std::string::String substring.

set ≈ a collection such as ::std::collections::BTreeSet.

bag ≈ a collection such as ::std::collections::BTreeMap which has key counts.

Forms

Forms for panic! versus Err()

The assert macros have three forms that you can use depending on your goals:

assert_gt!(a, b); // return () or panic!(…), for typical compile-time testing

debug_assert_gt!(a, b); // return () or panic!(…), for a non-optimized runtime

assert_gt_as_result!(a, b); // return Result Ok(()) or Err(…), for any runtime

Forms for messages

The assert macros have forms for default messages versus custom messages.

assert_gt!(1, 2); // panic!("assertion failed: assert_gt(1, 2)…")

assert_gt!(1, 2, "message"); // panic!("message")

Forms for comparing an other versus an expression

Some assert macros have forms for comparing an other versus an expression:

assert_io_read_to_string_eq!(reader1, reader2); // reader1.read_to_string() = reader2.read_to_string()

assert_io_read_to_string_eq_expr!(reader, expr); // reader1.read_to_string() = expr

Changes summary

Version 8.x top changes

Add assert_fs_read_to_string_* macros for comparing files.

Rename assert_read_to_string_* macros to assert_io_read_to_string_*. If you use these macros, then please update your code to use the new naming convention.

Version 7.x top changes

Add assert_in_delta, assert_in_epsilon.

Add assert_fn_* macros with multiple arities.

Add cargo release for optimized tagged releases.

Version 6.x top changes

Add assert_starts_with, assert_ends_with, assert_contains, assert_is_match.

Add debug_assert_* macros everywhere.

Add GPL-3.0 license.

Tracking

Package: assertables-rust-crate

Version: 8.0.0

Created: 2021-03-30T15:47:49Z

Updated: 2024-09-02T15:45:13Z

License: MIT or Apache-2.0 or GPL-2.0 or GPL-3.0 or contact us for more

Contact: Joel Parker Henderson (joel@sixarm.com)

Modules

assert

assert_bag_eq

assert_bag_ne

assert_bag_subbag

assert_bag_superbag

assert_command_stderr_contains

assert_command_stderr_eq

assert_command_stderr_eq_expr

assert_command_stderr_is_match

assert_command_stdout_contains

assert_command_stdout_eq

assert_command_stdout_eq_expr

assert_command_stdout_is_match

assert_contains

assert_ends_with

assert_eq

assert_fn_eq

assert_fn_eq_expr

assert_fn_err_eq

assert_fn_err_eq_expr

assert_fn_err_ge

assert_fn_err_ge_expr

assert_fn_err_gt

assert_fn_err_gt_expr

assert_fn_err_le

assert_fn_err_le_expr

assert_fn_err_lt

assert_fn_err_lt_expr

assert_fn_err_ne

assert_fn_err_ne_expr

assert_fn_ge

assert_fn_ge_expr

assert_fn_gt

assert_fn_gt_expr

assert_fn_le

assert_fn_le_expr

assert_fn_lt

assert_fn_lt_expr

assert_fn_ne

assert_fn_ne_expr

assert_fn_ok_eq

assert_fn_ok_eq_expr

assert_fn_ok_ge

assert_fn_ok_ge_expr

assert_fn_ok_gt

assert_fn_ok_gt_expr

assert_fn_ok_le

assert_fn_ok_le_expr

assert_fn_ok_lt

assert_fn_ok_lt_expr

assert_fn_ok_ne

assert_fn_ok_ne_expr

assert_fs_read_to_string_contains

assert_fs_read_to_string_eq

assert_fs_read_to_string_eq_expr

assert_fs_read_to_string_ge

assert_fs_read_to_string_ge_expr

assert_fs_read_to_string_gt

assert_fs_read_to_string_gt_expr

assert_fs_read_to_string_le

assert_fs_read_to_string_le_expr

assert_fs_read_to_string_lt

assert_fs_read_to_string_lt_expr

assert_fs_read_to_string_matches

assert_fs_read_to_string_ne

assert_fs_read_to_string_ne_expr

assert_ge

assert_gt

assert_in_delta

assert_in_epsilon

assert_io_read_to_string_contains

assert_io_read_to_string_eq

assert_io_read_to_string_eq_expr

assert_io_read_to_string_ge

assert_io_read_to_string_ge_expr

assert_io_read_to_string_gt

assert_io_read_to_string_gt_expr

assert_io_read_to_string_le

assert_io_read_to_string_le_expr

assert_io_read_to_string_lt

assert_io_read_to_string_lt_expr

assert_io_read_to_string_matches

assert_io_read_to_string_ne

assert_io_read_to_string_ne_expr

assert_is_match

assert_le

assert_lt

assert_ne

assert_not_contains

assert_not_ends_with

assert_not_match

assert_not_starts_with

assert_program_args_stderr_contains

assert_program_args_stderr_eq

assert_program_args_stderr_eq_expr

assert_program_args_stderr_ge

assert_program_args_stderr_ge_expr

assert_program_args_stderr_gt

assert_program_args_stderr_gt_expr

assert_program_args_stderr_is_match

assert_program_args_stderr_le

assert_program_args_stderr_le_expr

assert_program_args_stderr_lt

assert_program_args_stderr_lt_expr

assert_program_args_stderr_ne

assert_program_args_stderr_ne_expr

assert_program_args_stdout_contains

assert_program_args_stdout_eq

assert_program_args_stdout_eq_expr

assert_program_args_stdout_ge

assert_program_args_stdout_ge_expr

assert_program_args_stdout_gt

assert_program_args_stdout_gt_expr

assert_program_args_stdout_is_match

assert_program_args_stdout_le

assert_program_args_stdout_le_expr

assert_program_args_stdout_lt

assert_program_args_stdout_lt_expr

assert_program_args_stdout_ne

assert_program_args_stdout_ne_expr

assert_set_disjoint

assert_set_eq

assert_set_joint

assert_set_ne

assert_set_subset

assert_set_superset

assert_starts_with

Macros

assert_as_result

Assert a condition is true.

assert_bag_eq

Assert a bag is equal to another.

assert_bag_eq_as_result

Assert a bag is equal to another.

assert_bag_ne

Assert a bag is not equal to another.

assert_bag_ne_as_result

Assert a bag is not equal to another.

assert_bag_subbag

Assert a bag is a subbag of another.

assert_bag_subbag_as_result

Assert a bag is a subbag of another.

assert_bag_superbag

Assert a bag is a superbag of another.

assert_bag_superbag_as_result

Assert a bag is a superbag of another.

assert_command_stderr_contains

Assert a command stderr string contains a given containee.

assert_command_stderr_contains_as_result

Assert a command stderr string contains a given containee.

assert_command_stderr_eq

Assert a command stderr string is equal to another.

assert_command_stderr_eq_as_result

Assert a command stderr string is equal to another.

assert_command_stderr_eq_expr

Assert a command stderr string is equal to an expression.

assert_command_stderr_eq_expr_as_result

Assert a command stderr string is equal to an expression.

assert_command_stderr_is_match

Assert a command stderr string is a match to a regex.

assert_command_stderr_is_match_as_result

Assert a command stderr string is a match to a regex.

assert_command_stdout_contains

Assert a command stdout string contains a given containee.

assert_command_stdout_contains_as_result

Assert a command stdout string contains a given containee.

assert_command_stdout_eq

Assert a command stdout string is equal to another.

assert_command_stdout_eq_as_result

Assert a command stdout string is equal to another.

assert_command_stdout_eq_expr

Assert a command stdout string is equal to an expression.

assert_command_stdout_eq_expr_as_result

Assert a command stdout string is equal to an expression.

assert_command_stdout_is_match

Assert a command stdout string is a match to a regex.

assert_command_stdout_is_match_as_result

Assert a command stdout string is a match to a regex.

assert_contains

Assert a container is a match for an expression.

assert_contains_as_result

Assert an expression (such as a string) contains an expression (such as a substring).

assert_ends_with

Assert an expression (such as a string) ends with an expression (such as a string).

assert_ends_with_as_result

Assert an expression (such as a string) ends with an expression (such as a substring).

assert_eq_as_result

Assert an expression is equal to another.

assert_fn_eq

Assert a function output is equal to another function output.

assert_fn_eq_as_result

Assert a function output is equal to another function output.

assert_fn_eq_expr

Assert a function output is equal to an expression.

assert_fn_eq_expr_as_result

Assert a function output is equal to an expression.

assert_fn_err_eq

Assert a function err() is equal to another.

assert_fn_err_eq_as_result

Assert a function err() is equal to another.

assert_fn_err_eq_expr

Assert a function err() is equal to an expression.

assert_fn_err_eq_expr_as_result

Assert a function err() is equal to an expression.

assert_fn_err_ge

Assert a function err() is greater than or equal to another.

assert_fn_err_ge_as_result

Assert a function err() is greater than or equal to another.

assert_fn_err_ge_expr

Assert a function err() is greater than or equal to an expression.

assert_fn_err_ge_expr_as_result

Assert a function err() is greater than or equal to an expression.

assert_fn_err_gt

Assert a function err() is greater than another.

assert_fn_err_gt_as_result

Assert a function err() is greater than another.

assert_fn_err_gt_expr

Assert a function err() is greater than an expression.

assert_fn_err_gt_expr_as_result

Assert a function err() is greater than an expression.

assert_fn_err_le

Assert a function err() is less than or equal to another.

assert_fn_err_le_as_result

Assert a function err() is less than or equal to another.

assert_fn_err_le_expr

Assert a function err() is less than or equal to an expression.

assert_fn_err_le_expr_as_result

Assert a function err() is less than or equal to an expression.

assert_fn_err_lt

Assert a function err() is less than another.

assert_fn_err_lt_as_result

Assert a function err() is less than another.

assert_fn_err_lt_expr

Assert a function err() is less than an expression.

assert_fn_err_lt_expr_as_result

Assert a function err() is less than an expression.

assert_fn_err_ne

Assert a function err() is not equal to another.

assert_fn_err_ne_as_result

Assert a function err() is not equal to another.

assert_fn_err_ne_expr

Assert a function err() is not equal to an expression.

assert_fn_err_ne_expr_as_result

Assert a function err() is not equal to an expression.

assert_fn_ge

Assert a function output is greater than or equal to another.

assert_fn_ge_as_result

Assert a function output is greater than or equal to another.

assert_fn_ge_expr

Assert a function output is greater than or equal to an expression.

assert_fn_ge_expr_as_result

Assert a function output is greater than or equal to an expression.

assert_fn_gt

Assert a function output is greater than another.

assert_fn_gt_as_result

Assert a function output is greater than another.

assert_fn_gt_expr

Assert a function output is greater than an expression.

assert_fn_gt_expr_as_result

Assert a function output is greater than an expression.

assert_fn_le

Assert a function output is less than or equal to another.

assert_fn_le_as_result

Assert a function output is less than or equal to another.

assert_fn_le_expr

Assert a function output is less than or equal to an expression.

assert_fn_le_expr_as_result

Assert a function output is less than or equal to an expression.

assert_fn_lt

Assert a function output is less than another.

assert_fn_lt_as_result

Assert a function output is less than another.

assert_fn_lt_expr

Assert a function output is less than an expression.

assert_fn_lt_expr_as_result

Assert a function output is less than an expression.

assert_fn_ne

Assert a function output is not equal to another.

assert_fn_ne_as_result

Assert a function output is not equal to another.

assert_fn_ne_expr

Assert a function output is not equal to an expression.

assert_fn_ne_expr_as_result

Assert a function output is not equal to an expression.

assert_fn_ok_eq

Assert a function ok() is equal to another.

assert_fn_ok_eq_as_result

Assert a function ok() is equal to another.

assert_fn_ok_eq_expr

Assert a function ok() is equal to an expression.

assert_fn_ok_eq_expr_as_result

Assert a function ok() is equal to an expression.

assert_fn_ok_ge

Assert a function ok() is greater than or equal to another.

assert_fn_ok_ge_as_result

Assert a function ok() is greater than or equal to another.

assert_fn_ok_ge_expr

Assert a function ok() is greater than or equal to an expression.

assert_fn_ok_ge_expr_as_result

Assert a function ok() is greater than or equal to an expression.

assert_fn_ok_gt

Assert a function ok() is greater than another.

assert_fn_ok_gt_as_result

Assert a function ok() is greater than another.

assert_fn_ok_gt_expr

Assert a function ok() is greater than an expression.

assert_fn_ok_gt_expr_as_result

Assert a function ok() is greater than an expression.

assert_fn_ok_le

Assert a function ok() is less than or equal to another.

assert_fn_ok_le_as_result

Assert a function ok() is less than or equal to another.

assert_fn_ok_le_expr

Assert a function ok() is less than or equal to an expression.

assert_fn_ok_le_expr_as_result

Assert a function ok() is less than or equal to an expression.

assert_fn_ok_lt

Assert a function ok() is less than another.

assert_fn_ok_lt_as_result

Assert a function ok() is less than another.

assert_fn_ok_lt_expr

Assert a function ok() is less than an expression.

assert_fn_ok_lt_expr_as_result

Assert a function ok() is less than an expression.

assert_fn_ok_ne

Assert a function ok() is not equal to another.

assert_fn_ok_ne_as_result

Assert a function ok() is not equal to another.

assert_fn_ok_ne_expr

Assert a function ok() is not equal to an expression.

assert_fn_ok_ne_expr_as_result

Assert a function ok() is not equal to an expression.

assert_fs_read_to_string_contains

Assert a std::fs::read_to_string() contains a pattern.

assert_fs_read_to_string_contains_as_result

Assert a std::fs::read_to_string() contains a pattern.

assert_fs_read_to_string_eq

Assert a std::fs::read_to_string() value is equal to another.

assert_fs_read_to_string_eq_as_result

Assert a std::fs::read_to_string() is equal to another.

assert_fs_read_to_string_eq_expr

Assert a std::fs::read_to_string() value is equal to an expression.

assert_fs_read_to_string_eq_expr_as_result

Assert a std::fs::read_to_string() value is equal to an expression.

assert_fs_read_to_string_ge

Assert a std::fs::read_to_string() value is greater than or equal to another.

assert_fs_read_to_string_ge_as_result

Assert a std::fs::read_to_string() value is greater than or equal to another.

assert_fs_read_to_string_ge_expr

Assert a std::fs::read_to_string() value is greater than or equal to an expression.

assert_fs_read_to_string_ge_expr_as_result

Assert a std::fs::read_to_string() value is greater than or equal to an expression.

assert_fs_read_to_string_gt

Assert a std::fs::read_to_string() value is greater than another.

assert_fs_read_to_string_gt_as_result

Assert a std::fs::read_to_string() value is greater than another.

assert_fs_read_to_string_gt_expr

Assert a std::fs::read_to_string() value is greater than an expression.

assert_fs_read_to_string_gt_expr_as_result

Assert a std::fs::read_to_string() value is greater than an expression.

assert_fs_read_to_string_le

Assert a std::fs::read_to_string() value is less than or equal to another.

assert_fs_read_to_string_le_as_result

Assert a std::fs::read_to_string() value is less than or equal to another.

assert_fs_read_to_string_le_expr

Assert a std::fs::read_to_string() value is less than or equal to an expression.

assert_fs_read_to_string_le_expr_as_result

Assert a std::fs::read_to_string() value is less than or equal to an expression.

assert_fs_read_to_string_lt

Assert a std::fs::read_to_string() value is less than another.

assert_fs_read_to_string_lt_as_result

Assert a std::fs::read_to_string() value is less than another.

assert_fs_read_to_string_lt_expr

Assert a std::fs::read_to_string() value is less than an expression.

assert_fs_read_to_string_lt_expr_as_result

Assert a std::fs::read_to_string() value is less than an expression.

assert_fs_read_to_string_matches

Assert a std::fs::read_to_string() is a match to a regex.

assert_fs_read_to_string_matches_as_result

Assert a std::fs::read_to_string() is a match to a regex.

assert_fs_read_to_string_ne

Assert a std::fs::read_to_string() is not equal to another.

assert_fs_read_to_string_ne_as_result

Assert a std::fs::read_to_string() is not equal to another.

assert_fs_read_to_string_ne_expr

Assert a std::fs::read_to_string() is not equal to an expression.

assert_fs_read_to_string_ne_expr_as_result

Assert a std::fs::read_to_string() is not equal to an expression.

assert_ge

Assert a value is greater than or equal to an expression.

assert_ge_as_result

Assert a value is greater than or equal to an expression.

assert_gt

Assert a value is greater than an expression.

assert_gt_as_result

Assert a value is greater than an expression.

assert_in_delta

Assert a number is within delta of another number.

assert_in_delta_as_result

Assert a number is within delta of another number.

assert_in_epsilon

Assert a number is within epsilon of another number.

assert_in_epsilon_as_result

Assert a number is within epsilon of another number.

assert_io_read_to_string_contains

Assert a std::io::Read read_to_string() contains a pattern.

assert_io_read_to_string_contains_as_result

Assert a std::io::Read read_to_string() contains a pattern.

assert_io_read_to_string_eq

Assert a std::io::Read read_to_string() value is equal to another.

assert_io_read_to_string_eq_as_result

Assert a std::io::Read read_to_string() is equal to another.

assert_io_read_to_string_eq_expr

Assert a std::io::Read read_to_string() value is equal to an expression.

assert_io_read_to_string_eq_expr_as_result

Assert a std::io::Read read_to_string() value is equal to an expression.

assert_io_read_to_string_ge

Assert a std::io::Read read_to_string() value is greater than or equal to another.

assert_io_read_to_string_ge_as_result

Assert a std::io::Read read_to_string() value is greater than or equal to another.

assert_io_read_to_string_ge_expr

Assert a std::io::Read read_to_string() value is greater than or equal to an expression.

assert_io_read_to_string_ge_expr_as_result

Assert a std::io::Read read_to_string() value is greater than or equal to an expression.

assert_io_read_to_string_gt

Assert a std::io::Read read_to_string() value is greater than another.

assert_io_read_to_string_gt_as_result

Assert a std::io::Read read_to_string() value is greater than another.

assert_io_read_to_string_gt_expr

Assert a std::io::Read read_to_string() value is greater than an expression.

assert_io_read_to_string_gt_expr_as_result

Assert a std::io::Read read_to_string() value is greater than an expression.

assert_io_read_to_string_le

Assert a std::io::Read read_to_string() value is less than or equal to another.

assert_io_read_to_string_le_as_result

Assert a std::io::Read read_to_string() value is less than or equal to another.

assert_io_read_to_string_le_expr

Assert a std::io::Read read_to_string() value is less than or equal to an expression.

assert_io_read_to_string_le_expr_as_result

Assert a std::io::Read read_to_string() value is less than or equal to an expression.

assert_io_read_to_string_lt

Assert a std::io::Read read_to_string() value is less than another.

assert_io_read_to_string_lt_as_result

Assert a std::io::Read read_to_string() value is less than another.

assert_io_read_to_string_lt_expr

Assert a std::io::Read read_to_string() value is less than an expression.

assert_io_read_to_string_lt_expr_as_result

Assert a std::io::Read read_to_string() value is less than an expression.

assert_io_read_to_string_matches

Assert a std::io::Read read_to_string() is a match to a regex.

assert_io_read_to_string_matches_as_result

Assert a std::io::Read read_to_string() is a match to a regex.

assert_io_read_to_string_ne

Assert a std::io::Read read_to_string() is not equal to another.

assert_io_read_to_string_ne_as_result

Assert a std::io::Read read_to_string() is not equal to another.

assert_io_read_to_string_ne_expr

Assert a std::io::Read read_to_string() is not equal to an expression.

assert_io_read_to_string_ne_expr_as_result

Assert a std::io::Read read_to_string() is not equal to an expression.

assert_is_match

Assert a matcher is a match for an expression.

assert_is_match_as_result

Assert an expression (such as a regex) is a match for an expression (such as a string).

assert_le

Assert a value is less than or equal to an expression.

assert_le_as_result

Assert a value is less than or equal to an expression.

assert_lt

Assert a value is less than an expression.

assert_lt_as_result

Assert a value is less than an expression.

assert_ne_as_result

Assert an expression is not equal to an expression.

assert_not_contains

Assert a container is a match for an expression.

assert_not_contains_as_result

Assert an expression (such as a string) does not contain an expression (such as a substring).

assert_not_ends_with

Assert an expression (such as a string) does not end with an expression (such as a string).

assert_not_ends_with_as_result

Assert an expression (such as a string) does not end with an expression (such as a substring).

assert_not_match

Assert a matcher is a match for an expression.

assert_not_match_as_result

Assert an expression (such as a regex) is not a match for an expression (such as a string).

assert_not_starts_with

Assert an expression (such as a string) does not start with an expression (such as a string).

assert_not_starts_with_as_result

Assert an expression (such as a string) does not start with an expression (such as a substring).

assert_program_args_stderr_contains

Assert a command (built with program and args) stderr string contains a given containee.

assert_program_args_stderr_contains_as_result

Assert a command (built with program and args) stderr string contains a given containee.

assert_program_args_stderr_eq

Assert a command (built with program and args) stderr string is equal to another.

assert_program_args_stderr_eq_as_result

Assert a command (built with program and args) stderr string is equal to another.

assert_program_args_stderr_eq_expr

Assert a command (built with program and args) stderr string is equal to an expression.

assert_program_args_stderr_eq_expr_as_result

Assert a command (built with program and args) stderr string is equal to an expression.

assert_program_args_stderr_ge

Assert a command (built with program and args) stderr string is greater than or equal to another.

assert_program_args_stderr_ge_as_result

Assert a command (built with program and args) stderr string is greater than or equal to another.

assert_program_args_stderr_ge_expr

Assert a command (built with program and args) stderr string is greater than or equal to an expression.

assert_program_args_stderr_ge_expr_as_result

Assert a command (built with program and args) stderr string is greater than or equal to an expression.

assert_program_args_stderr_gt

Assert a command (built with program and args) stderr string is greater than to another.

assert_program_args_stderr_gt_as_result

Assert a command (built with program and args) stderr string is greater than another.

assert_program_args_stderr_gt_expr

Assert a command (built with program and args) stderr string is greater than an expression.

assert_program_args_stderr_gt_expr_as_result

Assert a command (built with program and args) stderr string is greater than an expression.

assert_program_args_stderr_is_match

Assert a command (built with program and args) stderr string is a match to a regex.

assert_program_args_stderr_is_match_as_result

Assert a command (built with program and args) stderr string is a match to a regex.

assert_program_args_stderr_le

Assert a command (built with program and args) stderr string is less than or equal to another.

assert_program_args_stderr_le_as_result

Assert a command (built with program and args) stderr string is less than or equal to another.

assert_program_args_stderr_le_expr

Assert a command (built with program and args) stderr string is less than or equal to an expression.

assert_program_args_stderr_le_expr_as_result

Assert a command (built with program and args) stderr string is less than or equal to an expression.

assert_program_args_stderr_lt

Assert a command (built with program and args) stderr string is less than another.

assert_program_args_stderr_lt_as_result

Assert a command (built with program and args) stderr string is less than another.

assert_program_args_stderr_lt_expr

Assert a command (built with program and args) stderr string is less than an expression.

assert_program_args_stderr_lt_expr_as_result

Assert a command (built with program and args) stderr string is less than an expression.

assert_program_args_stderr_ne

Assert a command (built with program and args) stderr string is not equal to another.

assert_program_args_stderr_ne_as_result

Assert a command (built with program and args) stderr string is not equal to another.

assert_program_args_stderr_ne_expr

Assert a command (built with program and args) stderr string is not equal to an expression.

assert_program_args_stderr_ne_expr_as_result

Assert a command (built with program and args) stderr string is not equal to an expression.

assert_program_args_stdout_contains

Assert a command (built with program and args) stdout string contains a given containee.

assert_program_args_stdout_contains_as_result

Assert a command (built with program and args) stdout string contains a given containee.

assert_program_args_stdout_eq

Assert a command (built with program and args) stdout string is equal to another.

assert_program_args_stdout_eq_as_result

Assert a command (built with program and args) stdout string is equal to another.

assert_program_args_stdout_eq_expr

Assert a command (built with program and args) stdout string is equal to an expression.

assert_program_args_stdout_eq_expr_as_result

Assert a command (built with program and args) stdout string is equal to an expression.

assert_program_args_stdout_ge

Assert a command (built with program and args) stdout string is greater than or equal to another.

assert_program_args_stdout_ge_as_result

Assert a command (built with program and args) stdout string is greater than or equal to another.

assert_program_args_stdout_ge_expr

Assert a command (built with program and args) stdout string is greater than or equal to an expression.

assert_program_args_stdout_ge_expr_as_result

Assert a command (built with program and args) stdout string is greater than or equal to an expression.

assert_program_args_stdout_gt

Assert a command (built with program and args) stdout string is greater than another.

assert_program_args_stdout_gt_as_result

Assert a command (built with program and args) stdout string is greater than to another.

assert_program_args_stdout_gt_expr

Assert a command (built with program and args) stdout string is greater than an expression.

assert_program_args_stdout_gt_expr_as_result

Assert a command (built with program and args) stdout string is greater than an expression.

assert_program_args_stdout_is_match

Assert a command (built with program and args) stdout string is a match to a regex.

assert_program_args_stdout_is_match_as_result

Assert a command (built with program and args) stdout string is a match to a regex.

assert_program_args_stdout_le

Assert a command (built with program and args) stdout string is less than or equal to another.

assert_program_args_stdout_le_as_result

Assert a command (built with program and args) stdout string is less than or equal to another.

assert_program_args_stdout_le_expr

Assert a command (built with program and args) stdout string is less than or equal to an expression.

assert_program_args_stdout_le_expr_as_result

Assert a command (built with program and args) stdout string is less than or equal to an expression.

assert_program_args_stdout_lt

Assert a command (built with program and args) stdout string is less than another.

assert_program_args_stdout_lt_as_result

Assert a command (built with program and args) stdout string is less than another.

assert_program_args_stdout_lt_expr

Assert a command (built with program and args) stdout string is less than an expression.

assert_program_args_stdout_lt_expr_as_result

Assert a command (built with program and args) stdout string is less than an expression.

assert_program_args_stdout_ne

Assert a command (built with program and args) stdout string is not equal to another.

assert_program_args_stdout_ne_as_result

Assert a command (built with program and args) stdout string is not equal to another.

assert_program_args_stdout_ne_expr

Assert a command (built with program and args) stdout string is not equal to an expression.

assert_program_args_stdout_ne_expr_as_result

Assert a command (built with program and args) stdout string is not equal to an expression.

assert_set_disjoint

Assert a set is disjoint with another.

assert_set_disjoint_as_result

Assert a set is disjoint with another.

assert_set_eq

Assert a set is equal to another.

assert_set_eq_as_result

Assert a set is equal to another.

assert_set_joint

Assert a set is joint with another.

assert_set_joint_as_result

Assert a set is joint with another.

assert_set_ne

Assert a set is not equal to another.

assert_set_ne_as_result

Assert a set is not equal to another.

assert_set_subset

Assert a set is a subset of another.

assert_set_subset_as_result

Assert a set is a subset of another.

assert_set_superset

Assert a set is a superset of another.

assert_set_superset_as_result

Assert a set is a superset of another.

assert_starts_with

Assert an expression (such as a string) starts with an expression (such as a string).

assert_starts_with_as_result

Assert an expression (such as a string) starts with an expression (such as a substring).

debug_assert_bag_eq

Assert a bag is equal to another.

debug_assert_bag_ne

Assert a bag is not equal to another.

debug_assert_bag_subbag

Assert a bag is a subbag of another.

debug_assert_bag_superbag

Assert a bag is a superbag of another.

debug_assert_command_stderr_contains

Assert a command stderr string contains a given containee.

debug_assert_command_stderr_eq

Assert a command stderr string is equal to another.

debug_assert_command_stderr_eq_expr

Assert a command stderr string is equal to an expression.

debug_assert_command_stderr_is_match

Assert a command stderr string is a match to a regex.

debug_assert_command_stdout_contains

Assert a command stdout string contains a given containee.

debug_assert_command_stdout_eq

Assert a command stdout string is equal to another.

debug_assert_command_stdout_eq_expr

Assert a command stdout string is equal to an expression.

debug_assert_command_stdout_is_match

Assert a command stdout string is a match to a regex.

debug_assert_contains

Assert a container is a match for an expression.

debug_assert_ends_with

Assert an expression (such as a string) ends with an expression (such as a string).

debug_assert_eq

Assert an expression is equal to an expression.

debug_assert_fn_eq

Assert a function output is equal to another function output.

debug_assert_fn_eq_expr

Assert a function output is equal to an expression.

debug_assert_fn_err_eq

Assert a function err() is equal to another.

debug_assert_fn_err_eq_expr

Assert a function err() is equal to an expression.

debug_assert_fn_err_ge

Assert a function err() is greater than or equal to another.

debug_assert_fn_err_ge_expr

Assert a function err() is greater than or equal to an expression.

debug_assert_fn_err_gt

Assert a function err() is greater than another.

debug_assert_fn_err_gt_expr

Assert a function err() is greater than an expression.

debug_assert_fn_err_le

Assert a function err() is less than or equal to another.

debug_assert_fn_err_le_expr

Assert a function err() is less than or equal to an expression.

debug_assert_fn_err_lt

Assert a function err() is less than another.

debug_assert_fn_err_lt_expr

Assert a function err() is less than an expression.

debug_assert_fn_err_ne

Assert a function err() is not equal to another.

debug_assert_fn_err_ne_expr

Assert a function err() is not equal to an expression.

debug_assert_fn_ge

Assert a function output is greater than or equal to another.

debug_assert_fn_ge_expr

Assert a function output is greater than or equal to an expression.

debug_assert_fn_gt

Assert a function output is greater than another.

debug_assert_fn_gt_expr

Assert a function output is greater than an expression.

debug_assert_fn_le

Assert a function output is less than or equal to another.

debug_assert_fn_le_expr

Assert a function output is less than or equal to an expression.

debug_assert_fn_lt

Assert a function output is less than another.

debug_assert_fn_lt_expr

Assert a function output is less than an expression.

debug_assert_fn_ne

Assert a function output is not equal to another.

debug_assert_fn_ne_expr

Assert a function output is not equal to an expression.

debug_assert_fn_ok_eq

Assert a function ok() is equal to another.

debug_assert_fn_ok_eq_expr

Assert a function ok() is equal to an expression.

debug_assert_fn_ok_ge

Assert a function ok() is greater than or equal to another.

debug_assert_fn_ok_ge_expr

Assert a function ok() is greater than or equal to an expression.

debug_assert_fn_ok_gt

Assert a function ok() is greater than another.

debug_assert_fn_ok_gt_expr

Assert a function ok() is greater than an expression.

debug_assert_fn_ok_le

Assert a function ok() is less than or equal to another.

debug_assert_fn_ok_le_expr

Assert a function ok() is less than or equal to an expression.

debug_assert_fn_ok_lt

Assert a function ok() is less than another.

debug_assert_fn_ok_lt_expr

Assert a function ok() is less than an expression.

debug_assert_fn_ok_ne

Assert a function ok() is not equal to another.

debug_assert_fn_ok_ne_expr

Assert a function ok() is not equal to an expression.

debug_assert_ge

Assert a value is greater than or equal to an expression.

debug_assert_gt

Assert a value is greater than an expression.

debug_assert_in_delta

Assert a number is within delta of another number.

debug_assert_in_epsilon

Assert a number is within epsilon of another number.

debug_assert_io_read_to_string_contains

Assert a std::io::Read read_to_string() contains a pattern.

debug_assert_io_read_to_string_eq

Assert a std::io::Read read_to_string() value is equal to another.

debug_assert_io_read_to_string_eq_expr

Assert a std::io::Read read_to_string() value is equal to an expression.

debug_assert_io_read_to_string_ge

Assert a std::io::Read read_to_string() value is greater than or equal to another.

debug_assert_io_read_to_string_ge_expr

Assert zzz.

debug_assert_io_read_to_string_gt

Assert a std::io::Read read_to_string() value is greater than another.

debug_assert_io_read_to_string_gt_expr

Assert a std::io::Read read_to_string() value is greater than an expression.

debug_assert_io_read_to_string_le

Assert a std::io::Read read_to_string() value is less than or equal to another.

debug_assert_io_read_to_string_le_expr

Assert a std::io::Read read_to_string() value is less than or equal to an expression.

debug_assert_io_read_to_string_lt

Assert a std::io::Read read_to_string() value is less than another.

debug_assert_io_read_to_string_lt_expr

Assert a std::io::Read read_to_string() value is less than an expression.

debug_assert_io_read_to_string_matches

Assert a std::io::Read read_to_string() is a match to a regex.

debug_assert_io_read_to_string_ne

Assert a std::io::Read read_to_string() is not equal to another.

debug_assert_io_read_to_string_ne_expr

Assert a std::io::Read read_to_string() is not equal to an expression.

debug_assert_is_match

Assert a matcher is a match for an expression.

debug_assert_le

Assert a value is less than or equal to an expression.

debug_assert_lt

Assert a value is less than an expression.

debug_assert_ne

Assert an expression is not equal to an expression.

debug_assert_not_contains

Assert a container is a match for an expression.

debug_assert_not_ends_with

Assert an expression (such as a string) does not end with an expression (such as a string).

debug_assert_not_match

Assert a matcher is a match for an expression.

debug_assert_not_starts_with

Assert an expression (such as a string) does not start with an expression (such as a string).

debug_assert_program_args_stderr_contains

Assert a command (built with program and args) stderr string contains a given containee.

debug_assert_program_args_stderr_eq

Assert a command (built with program and args) stderr string is equal to another.

debug_assert_program_args_stderr_eq_expr

Assert a command (built with program and args) stderr string is equal to an expression.

debug_assert_program_args_stderr_ge

Assert a command (built with program and args) stderr string greater than or equal to another.

debug_assert_program_args_stderr_ge_expr

Assert a command (built with program and args) stderr string is greater than or equal to an expression.

debug_assert_program_args_stderr_gt

Assert a command (built with program and args) stderr string is equal to another.

debug_assert_program_args_stderr_gt_expr

Assert a command (built with program and args) stderr string is greater than an expression.

debug_assert_program_args_stderr_is_match

Assert a command (built with program and args) stderr string is a match to a regex.

debug_assert_program_args_stderr_le

Assert a command (built with program and args) stderr string is less than or equal to another.

debug_assert_program_args_stderr_le_expr

Assert a command (built with program and args) stderr string is equal to an expression.

debug_assert_program_args_stderr_lt

Assert a command (built with program and args) stderr string is less than another.

debug_assert_program_args_stderr_lt_expr

Assert a command (built with program and args) stderr string is less than an expression.

debug_assert_program_args_stderr_ne

Assert a command (built with program and args) stderr string is not equal to another.

debug_assert_program_args_stderr_ne_expr

Assert a command (built with program and args) stderr string is not equal to an expression.

debug_assert_program_args_stdout_contains

Assert a command (built with program and args) stdout string contains a given containee.

debug_assert_program_args_stdout_eq

Assert a command (built with program and args) stdout string is equal to another.

debug_assert_program_args_stdout_eq_expr

Assert a command (built with program and args) stdout string is equal to an expression.

debug_assert_program_args_stdout_ge

Assert a command (built with program and args) stdout string is greater than or equal to another.

debug_assert_program_args_stdout_ge_expr

Assert a command (built with program and args) stdout string is greater than or equal to an expression.

debug_assert_program_args_stdout_gt

Assert a command (built with program and args) stdout string is greater than another.

debug_assert_program_args_stdout_gt_expr

Assert a command (built with program and args) stdout string is greater than an expression.

debug_assert_program_args_stdout_is_match

Assert a command (built with program and args) stdout string is a match to a regex.

debug_assert_program_args_stdout_le

Assert a command (built with program and args) stdout string is less than or equal to another.

debug_assert_program_args_stdout_le_expr

Assert a command (built with program and args) stdout string is less than or equal to an expression.

debug_assert_program_args_stdout_lt

Assert a command (built with program and args) stdout string is less than another.

debug_assert_program_args_stdout_lt_expr

Assert a command (built with program and args) stdout string is less than an expression.

debug_assert_program_args_stdout_ne

Assert a command (built with program and args) stdout string is not equal to another.

debug_assert_program_args_stdout_ne_expr

Assert a command (built with program and args) stdout string is not equal to an expression.

debug_assert_set_disjoint

Assert a set is disjoint with another.

debug_assert_set_eq

Assert a set is equal to another.

debug_assert_set_joint

Assert a set is joint with another.

debug_assert_set_ne

Assert a set is not equal to another.

debug_assert_set_subset

Assert a set is a subset of another.

debug_assert_set_superset

Assert a set is a superset of another.

debug_assert_starts_with

Assert an expression (such as a string) starts with an expression (such as a string).

debug_assert_zzz

Assert zzz.

debug_read_to_string_contains

Assert a std::fs::read_to_string() contains a pattern.

debug_read_to_string_eq

Assert a std::fs::read_to_string() value is equal to another.

debug_read_to_string_eq_expr

Assert a std::fs::read_to_string() value is equal to an expression.

debug_read_to_string_ge

Assert a std::fs::read_to_string() value is greater than or equal to another.

debug_read_to_string_ge_expr

Assert zzz.

debug_read_to_string_gt

Assert a std::fs::read_to_string() value is greater than another.

debug_read_to_string_gt_expr

Assert a std::fs::read_to_string() value is greater than an expression.

debug_read_to_string_le

Assert a std::fs::read_to_string() value is less than or equal to another.

debug_read_to_string_le_expr

Assert a std::fs::read_to_string() value is less than or equal to an expression.

debug_read_to_string_lt

Assert a std::fs::read_to_string() value is less than another.

debug_read_to_string_lt_expr

Assert a std::fs::read_to_string() value is less than an expression.

debug_read_to_string_matches

Assert a std::fs::read_to_string() is a match to a regex.

debug_read_to_string_ne

Assert a std::fs::read_to_string() is not equal to another.

debug_read_to_string_ne_expr

Assert a std::fs::read_to_string() is not equal to an expression.