Module lib::assert_program_args

Assert for comparing programs with arguments.

These macros help with calling external programs with arguments, then capturing the standard output stream and standard error stream.

These macros have corresponding macros in the module assert_command.

Program args stdout

Compare program and arguments standard output to another program and arguments standard output:

assert_program_args_stdout_eq2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stdout = command2 with program2 and args2 to stdout

assert_program_args_stdout_ne2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stdout ≠ command2 with program2 and args2 to stdout

assert_program_args_stdout_lt2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stdout < command2 with program2 and args2 to stdout

assert_program_args_stdout_le2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stdout ≤ command2 with program2 and args2 to stdout

assert_program_args_stdout_gt2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stdout > command2 with program2 and args2 to stdout

assert_program_args_stdout_ge2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stdout ≥ command2 with program2 and args2 to stdout

Compare program and arguments standard output to an expression:

assert_program_args_stdout_eq!(program, args, expr) ≈ command using program and args to stdout = expr

assert_program_args_stdout_ne!(program, args, expr) ≈ command using program and args to stdout ≠ expr

assert_program_args_stdout_lt!(program, args, expr) ≈ command using program and args to stdout < expr

assert_program_args_stdout_le!(program, args, expr) ≈ command using program and args to stdout ≤ expr

assert_program_args_stdout_gt!(program, args, expr) ≈ command using program and args to stdout > expr

assert_program_args_stdout_ge!(program, args, expr) ≈ command using program and args to stdout ≥ expr

Assert program and arguments standard output as a string:

assert_program_args_stdout_string_contains!(program, args, containee) ≈ command using program and args to stdout string contains containee

assert_program_args_stdout_string_is_match!(program, args, matcher) ≈ matcher is match with command using program and args

Program args stderr

Compare program and arguments standard error to another program and arguments standard error:

assert_program_args_stderr_eq2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stderr = command2 with program2 and args2 to stderr

assert_program_args_stderr_ne2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stderr ≠ command2 with program2 and args2 to stderr

assert_program_args_stderr_lt2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stderr < command2 with program2 and args2 to stderr

assert_program_args_stderr_le2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stderr ≤ command2 with program2 and args2 to stderr

assert_program_args_stderr_gt2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stderr > command2 with program2 and args2 to stderr

assert_program_args_stderr_ge2!(program1, args1, program2, args2) ≈ command using program1 and args1 to stderr ≥ command2 with program2 and args2 to stderr

Compare program and arguments standard error to an expression:

assert_program_args_stderr_eq!(program, args, expr) ≈ command using program and args to stderr = expr

assert_program_args_stderr_ne!(program, args, expr) ≈ command using program and args to stderr ≠ expr

assert_program_args_stderr_lt!(program, args, expr) ≈ command using program and args to stderr < expr

assert_program_args_stderr_le!(program, args, expr) ≈ command using program and args to stderr ≤ expr

assert_program_args_stderr_gt!(program, args, expr) ≈ command using program and args to stderr > expr

assert_program_args_stderr_ge!(program, args, expr) ≈ command using program and args to stderr ≥ expr

Assert program and arguments standard error as a string:

assert_program_args_stderr_string_contains!(program, args, containee) ≈ command using program and args to stderr string contains containee

assert_program_args_stderr_string_is_match!(program, args, matcher) ≈ matcher is match with command using program and args

Example

use assertables::*;

let a_program = "bin/printf-stdout";
let a_args = ["%s", "alfa"];
let b_program = "bin/printf-stdout";
let b_args = ["%s%s%s%s", "a", "l", "f", "a"];
assert_program_args_stdout_eq2!(&a_program, &a_args, &b_program, &b_args);

Modules

assert_program_args_stderr_contains

Assert a command (built with program and args) stderr string contains a given containee.

assert_program_args_stderr_eq

Assert a command (built with program and args) stderr string is equal to an expression.

assert_program_args_stderr_eq2

Assert a command (built with program and args) stderr string is equal to another.

assert_program_args_stderr_ge

Assert a command (built with program and args) stderr string is greater than or equal to an expression.

assert_program_args_stderr_ge2

Assert a command (built with program and args) stderr string is greater than or equal to another.

assert_program_args_stderr_gt

Assert a command (built with program and args) stderr string is greater than an expression.

assert_program_args_stderr_gt2

Assert a command (built with program and args) stderr string is greater than to another.

assert_program_args_stderr_is_match

Assert a command (built with program and args) stderr string is a match to a regex.

assert_program_args_stderr_le

Assert a command (built with program and args) stderr string is less than or equal to an expression.

assert_program_args_stderr_le2

Assert a command (built with program and args) stderr string is less than or equal to another.

assert_program_args_stderr_lt

Assert a command (built with program and args) stderr string is less than an expression.

assert_program_args_stderr_lt2

Assert a command (built with program and args) stderr string is less than another.

assert_program_args_stderr_ne

Assert a command (built with program and args) stderr string is not equal to an expression.

assert_program_args_stderr_ne2

Assert a command (built with program and args) stderr string is not equal to another.

assert_program_args_stderr_string_contains

Assert a command (built with program and args) stderr string contains a given containee.

assert_program_args_stderr_string_is_match

Assert a command (built with program and args) stderr string is a match to a regex.

assert_program_args_stdout_contains

Assert a command (built with program and args) stdout string contains a given containee.

assert_program_args_stdout_eq

Assert a command (built with program and args) stdout string is equal to an expression.

assert_program_args_stdout_eq2

Assert a command (built with program and args) stdout string is equal to another.

assert_program_args_stdout_ge

Assert a command (built with program and args) stdout string is greater than or equal to an expression.

assert_program_args_stdout_ge2

Assert a command (built with program and args) stdout string is greater than or equal to another.

assert_program_args_stdout_gt

Assert a command (built with program and args) stdout string is greater than an expression.

assert_program_args_stdout_gt2

Assert a command (built with program and args) stdout string is greater than another.

assert_program_args_stdout_is_match

Assert a command (built with program and args) stdout string is a match to a regex.

assert_program_args_stdout_le

Assert a command (built with program and args) stdout string is less than or equal to an expression.

assert_program_args_stdout_le2

Assert a command (built with program and args) stdout string is less than or equal to another.

assert_program_args_stdout_lt

Assert a command (built with program and args) stdout string is less than an expression.

assert_program_args_stdout_lt2

Assert a command (built with program and args) stdout string is less than another.

assert_program_args_stdout_ne

Assert a command (built with program and args) stdout string is not equal to an expression.

assert_program_args_stdout_ne2

Assert a command (built with program and args) stdout string is not equal to another.

assert_program_args_stdout_string_contains

Assert a command (built with program and args) stdout string contains a given containee.

assert_program_args_stdout_string_is_match

Assert a command (built with program and args) stdout string is a match to a regex.